% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adresse.switch.R
\name{adresse.switch}
\alias{adresse.switch}
\title{Switcher de localisation de fichier}
\arguments{
\item{url}{Adresse du fichier à ouvrir}
}
\description{
Cette fonction permet de retrouver automatique l'adresse d'un fichier entre l'ordi portable et l'ordi fixe
}
\examples{
adresse.switch("NAS-JB/Études/2014_Les Rousses/Résultats/Poissons/ResultatsRoussestotal-28janvier2015.xlsx")
adresse.switch("NAS-DATA/Poissons/PDPG/Stations PDPG_V1.xlsx")
fish <- read_excel(adresse.switch("NAS-JB/Études/2014_Les Rousses/Résultats/Poissons/ResultatsRoussestotal-28janvier2015.xlsx"), sheet = 2)
}
