% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDD.connexion.R
\name{BDD.connexion}
\alias{BDD.connexion}
\title{Chargement de la base de données Web-PDPG}
\usage{
BDD.connexion(
  dbname = NA_character_,
  hote = NA_character_,
  port = NA_character_,
  utilisateur = NA_character_,
  motdepasse = NA_character_
)
}
\arguments{
\item{dbname}{Nom de la base de données}

\item{hote}{Hôte de la base de données}

\item{port}{Port de la base de données}

\item{utilisateur}{Nom d'utilisateur}

\item{motdepasse}{Mot de passe du portefeuille}
}
\description{
Cette fonction permet de charger la connexion vers la base de données du Web-PDPG
}
\examples{
bd_w <- BDD.connexion(dbname, hote, port, utilisateur, motdepasse)
}
