% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDD.format.R
\name{BDD.format}
\alias{BDD.format}
\title{Mise au format des données pour ajout BDD}
\arguments{
\item{data}{Chronique à valider}
}
\description{
Cette fonction permet de formater les données pour les ajouter aux bases de données respectives
}
\examples{
BDD.format(data)
}
\keyword{data}
