% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDD.listetables.R
\name{BDD.listetables}
\alias{BDD.listetables}
\title{Chargement des tables et champs de nas-sig-data}
\usage{
BDD.listetables()
}
\description{
Cette fonction permet de charger les noms des tables et des champs de la base de donnéesnas-sig-data
}
\examples{
BDD.listetables()
BDD.listetables() \%>\% View()
}
