% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDD.SIG.R
\name{BDD.SIG}
\alias{BDD.SIG}
\title{Export BDD vers SIG des stations}
\usage{
BDD.SIG()
}
\arguments{
\item{data}{Chronique à valider}
}
\description{
Cette fonction permet d'exporter automatiquement les stations des différentes bases de données vers les couches SIG correspondantes
}
\examples{
BDD.SIG()
}
\keyword{data}
