% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.ajout.R
\name{chronique.ajout}
\alias{chronique.ajout}
\title{Ajout de données brutes de chronique}
\arguments{
\item{data}{Chronique à modifier}

\item{cUnite}{\code{°C} (par défault)}

\item{cTypeMesure}{\code{Thermie} (par défault)}

\item{cNvelleValeur}{Valeur à ajouter}

\item{cValidation}{\code{Validé} (par défault)}

\item{cModeAcquisition}{\code{Estimé} (par défault)}
}
\description{
Permet d'ajouter des données brutes de chronique présentant des valeurs aberrantes. ATTENTION : ne gère pas les ID
}
\examples{
chronique.ajout(data, cCodeRDT = "BONbaro", cCapteur = "P0352", cDate = "2014-12-27", cHeure = "19:00:00", cUnite = "°C", cTypeMesure = "Thermie", cNvelleValeur = 15, cValidation = "Validé", cModeAcquisition = "Estimé")
}
\keyword{chronique}
