% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.analyse.R
\name{chronique.analyse}
\alias{chronique.analyse}
\title{Analyse de chroniques}
\arguments{
\item{data}{Data.frame contenant a minima une colonne chmes_date, une colonne chmes_heure et une colonne chmes_valeur}

\item{typemesure}{Défini le type de données et modifie l'analyse en fonction}

\item{pasdetemps}{Indique le pas de temps entre chaque mesure : 1 pour une heure (1 par défaut)}

\item{Vmin}{Valeur inférieure de seuil d'analyse}

\item{Vmax}{Valeur supérieure de seuil d'analyse}

\item{VminExt}{Valeur inférieure extrême de seuil d'analyse}

\item{VmaxExt}{Valeur supérieure extrême de seuil d'analyse}

\item{DegresJours}{Calcul des degrés/jours (\code{FALSE} par défault)}
}
\description{
Cette fonction permet d'analyser des chroniques de mesures (température, niveaux, etc.)
}
\examples{
chronique.analyse("ANG0-4", 2016, data, pasdetemps = 1)
chronique.analyse(as.character(Stations[i,2]), Annee, data, pasdetemps = 1)
}
\keyword{chronique}
