% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.capteurs.R
\name{chronique.capteurs}
\alias{chronique.capteurs}
\title{Listage des capteurs de chroniques}
\arguments{
\item{Nom}{Nom recherché}

\item{Recherche}{Type de données recherchées}
}
\description{
Cette fonction permet de lister les capteurs de la BDD Chroniques
}
\examples{
chronique.capteurs("CD39","Propriétaire")
chronique.capteurs("Thermie","Type")
chronique.capteurs("Hobo UA-001-64","Modèle")
chronique.capteurs("10165890","Numéro")
chronique.capteurs("PDPG","Projet")
chronique.capteurs("OK","État")
}
\keyword{capteurs}
