% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.classescalendaires.R
\name{chronique.classescalendaires}
\alias{chronique.classescalendaires}
\title{Calcul de classes de valeurs de chroniques}
\usage{
chronique.classescalendaires(
  data = data,
  typemesure = c("Thermie", "Thermie barométrique", "Thermie piézométrique",
    "Barométrie", "Piézométrie", "Piézométrie brute", "Piézométrie compensée",
    "Piézométrie calée", "Piézométrie NGF", "Oxygénation", "Hydrologie",
    "Pluviométrie"),
  classe_variable = c("VMedJ", "VMinJ", "VMoyJ", "VMaxJ", "VAmpliJ", "VAmpliSigneJ",
    "VarJ", "NMesuresJ", "SommeMoyJ"),
  datedebutanneebiol = "10-01",
  datedebutanneeneutre = "10-01"
)
}
\arguments{
\item{data}{Data.frame issu de chronique.agregation (données journalières)}

\item{typemesure}{Défini le type de données et modifie les légendes en fonction. Ignoré si le champ chmes_typemesure est présent dans data}

\item{classe_variable}{Type de variable journalière agrégée à utiliser : \code{VMedJ} (par défaut), \code{VMinJ}, \code{VMoyJ}, \code{VMaxJ}, \code{VAmpliJ}, \code{VAmpliSigneJ}, \code{VarJ}, \code{NMesuresJ} ou \code{SommeMoyJ}}

\item{datedebutanneebiol}{Date de démarrage de l'année biologique : 10-01 (par défaut - 1er octobre), pour l'affichage sous forme de courbes}

\item{datedebutanneeneutre}{Date de démarrage de l'année neutre : 10-01 (par défaut - 1er octobre), pour l'affichage sous forme d'années neutres}
}
\description{
Cette fonction permet de calculer les classes de valeurs de référence pour des chroniques de mesures (température, niveaux, etc.).
}
\examples{
data \%>\% chronique.agregation() \%>\% purrr::pluck(2) \%>\% chronique.classescalendaires()
data \%>\% chronique.agregation(instantanne = F, mensuel = F, annuel = F, integral = F) \%>\% chronique.classescalendaires()
donnees_journalieres \%>\% chronique.classescalendaires(classe_variable = "VMinJ")
donnees_journalieres \%>\% chronique.classescalendaires(classe_variable = "VMaxJ")
}
\keyword{chronique}
