% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.complete.R
\name{chronique.complete}
\alias{chronique.complete}
\title{Complément de chroniques}
\usage{
chronique.complete(data = data)
}
\arguments{
\item{data}{Data.frame d'une série de mesures de chroniques au format \code(mesures_structure)}
}
\description{
Cette fonction permet de compléter les chroniques présentant des valeurs manquantes (une station avec un unique paramètre)
}
\examples{
chronique.complete(data)
data \%>\% group_split(chmes_coderhj) \%>\% map(~ chronique.complete(.)) \%>\% list_rbind() 
}
\keyword{chronique}
