% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.contexte.R
\name{chronique.contexte}
\alias{chronique.contexte}
\title{Contexte de données de chroniques}
\usage{
chronique.contexte(data = data)
}
\arguments{
\item{data}{Data.frame issu des fonctions chronique.mesures, chronique.agregation ou chronique.resultats}
}
\description{
Cette fonction permet d'établir le contexte de données de chroniques (nombre de stations, d'années biologiques, de types de mesures, etc.)
}
\examples{
chronique.mesures("DRO6-8") \%>\% chronique.contexte()
Contexte <- chronique.contexte(Mesures)
Contexte <- chronique.contexte(chronique.mesures("DRO6-8"))
}
\keyword{chronique}
