% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.correction.R
\name{chronique.correction}
\alias{chronique.correction}
\title{Correction de données brutes de chronique}
\usage{
chronique.correction(data, cCodeRDT = "BONbaro", cCapteur = "P0352",
  cDate = "2014-12-27", cHeure = "19:00:00", cValeur = 985.3,
  cTypeMesure = "Piézométrie", cNvelleValeur = 700)
}
\arguments{
\item{data}{Chronique à modifier}

\item{cNvelleValeur}{Valeur de remplacement}
}
\description{
Permet de corriger des données brutes de chronique présentant des valeurs aberrantes. ATTENTION : ne gère pas les ID
}
\examples{
chronique.correction(data, cCodeRDT = "BONbaro", cCapteur = "P0352", cDate = "2014-12-27", cHeure = "19:00:00", cValeur = 985.3, cTypeMesure = "Piézométrie", cNvelleValeur = 700)
}
\keyword{chronique}

