% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.degresjours.R
\name{chronique.degresjours}
\alias{chronique.degresjours}
\title{Calcul d'une durée de degrés-jours}
\usage{
chronique.degresjours(
  data = data,
  espece = NA_character_,
  valeurdeclenchement = 10,
  typeseuildeclenchement = c("Montant", "Descendant"),
  nbjourssuccessifsdeclenchement = 2,
  datedeclenchementmanuelle = NA_character_,
  valeurdegresjours = 180
)
}
\arguments{
\item{data}{Data.frame de données journalières issu de la fonction \code{chronique.agregation}}

\item{espece}{Espèce recherchée (vide par défaut - Si complété, va directement chercher les valeurs de référence dans \code{data(PoissonsThermieReproduction)} - Exemple : TRF, BRO)}

\item{valeurdeclenchement}{Température de déclenchement de la phase de reproduction (pour recherche automatique de la date de déclenchement)}

\item{nbjourssuccessifsdeclenchement}{Nombre de jours successifs (\code{2} par défaut) devant afficher une température moyenne supérieure ou inférieure (en fonction de typevaleurdeclenchement) à la valeurdeclenchement}

\item{datedeclenchementmanuelle}{Si l'opérateur souhaite forcer manuellement la date de démarrage du calcul (au format 2020-12-01)}

\item{valeurdegresjours}{Valeur de degrés-jours de référence}

\item{typevaleurdeclenchement}{Si \code{Montant} (par défault), s'appuie sur une chronique printanière (valeurs croissantes : BRO, OBR, GAR, etc.). Si \code{Descendant}, s'appuie sur une chronique automnale (TRF)}
}
\description{
Cette fonction permet de calcul une date de démarrage et de fin de reproduction à partir de données de degrés-jours
}
\examples{
Resultats \%>\% chronique.degresjours()
Resultats \%>\% chronique.degresjours(valeurdeclenchement = 10, typeseuildeclenchement = "Montant", nbjourssuccessifsdeclenchement = 2, valeurdegresjours = 180)
Resultats \%>\% chronique.degresjours(valeurdegresjours = 180, datedeclenchementmanuelle = "2021-03-30")
test <- ValJours \%>\% filter(chmes_coderhj == "AIN22-2") \%>\% formatage.annee.biologique() \%>\% group_by(chmes_coderhj, chmes_anneebiol) \%>\% group_split() 
listeespecesatraiter <- c("TRF", "OBR", "BRO", "GAR", "TAN", "CCO", "HOT", "BAF")
argList <- list(x = test, y = listeespecesatraiter)
crossArg <- cross_df(argList)
test <- map2_dfr(crossArg$x, crossArg$y, chronique.degresjours)
}
\keyword{chronique}
