% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.figure.classescalendaires.R
\name{chronique.figure.classescalendaires}
\alias{chronique.figure.classescalendaires}
\title{Classes de valeurs de chroniques sous forme calendaire}
\usage{
chronique.figure.classescalendaires(
  data = data,
  Titre = "",
  typemesure = c("Thermie", "Thermie barométrique", "Thermie piézométrique",
    "Barométrie", "Piézométrie", "Piézométrie brute", "Piézométrie compensée",
    "Piézométrie calée", "Piézométrie NGF", "Oxygénation", "Hydrologie",
    "Pluviométrie"),
  classe_variable = c("VMedJ", "VMinJ", "VMoyJ", "VMaxJ", "VAmpliJ", "VAmpliSigneJ",
    "VarJ", "NMesuresJ", "SommeMoyJ"),
  stations = NA_character_,
  affichagevide = FALSE,
  datedebutanneebiol = "10-01",
  datedebutanneeneutre = "10-01",
  origine_donnees = NA_character_,
  save = F,
  projet = NA_character_,
  format = ".png"
)
}
\arguments{
\item{data}{Data.frame issu de chronique.agregation (données journalières)}

\item{Titre}{Titre du graphique (vide par défaut)}

\item{typemesure}{Défini le type de données et modifie les légendes en fonction. Ignoré si le champ chmes_typemesure est présent dans data}

\item{classe_variable}{Type de variable journalière agrégée à utiliser : \code{VMedJ} (par défaut), \code{VMinJ}, \code{VMoyJ}, \code{VMaxJ}, \code{VAmpliJ}, \code{VAmpliSigneJ}, \code{VarJ}, \code{NMesuresJ} ou \code{SommeMoyJ}}

\item{stations}{Dataframe contenant les stations, avec au minimum \code{chsta_coderhj} et \code{chsta_distancesource}}

\item{affichagevide}{Si \code{TRUE}, fait apparaître les années ne contenant pas de résultats. Si \code{FALSE} (par défault), ne fait apparaître que les années contenants des résultats}

\item{datedebutanneebiol}{Date de démarrage de l'année biologique : 10-01 (par défaut - 1er octobre), pour l'affichage sous forme de courbes}

\item{datedebutanneeneutre}{Date de démarrage de l'année neutre : 10-01 (par défaut - 1er octobre), pour l'affichage sous forme d'années neutres}

\item{origine_donnees}{Éventuelle source des données à afficher sur la figure (vide par défaut)}

\item{save}{Si \code{FALSE} (par défault), n'enregistre pas les figures. Si \code{TRUE}, les enregistre.}

\item{projet}{Nom du projet}

\item{format}{Défini le format d'enregistrement (par défaut .png)}
}
\description{
Cette fonction permet de représenter les classes de valeurs de référence pour des chroniques de mesures (température, niveaux, etc.) sous forme calendaire.
}
\examples{
data \%>\% chronique.agregation() \%>\% purrr::pluck(2) \%>\% chronique.figure.classescalendaires()
data \%>\% chronique.agregation(instantanne = F, mensuel = F, annuel = F, integral = F) \%>\% chronique.figure.classescalendaires()
donnees_journalieres \%>\% chronique.figure.classescalendaires(classe_variable = "VMinJ")
donnees_journalieres \%>\% chronique.figure.classescalendaires(classe_variable = "VMaxJ", origine_donnees = "PNRHJ")
}
\keyword{chronique}
