% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.figure.classesdensite.R
\name{chronique.figure.classesdensite}
\alias{chronique.figure.classesdensite}
\title{Classes de valeurs de chroniques sous forme de densités}
\usage{
chronique.figure.classesdensite(
  data = data,
  titre = "",
  origine_donnees = "",
  typemesure = c("Thermie", "Thermie barométrique", "Thermie piézométrique",
    "Barométrie", "Piézométrie", "Piézométrie brute", "Piézométrie compensée",
    "Piézométrie calée", "Piézométrie NGF", "Oxygénation", "Hydrologie",
    "Pluviométrie"),
  stations = NA_character_,
  affichagevide = FALSE,
  datedebutanneebiol = "10-01",
  datedebutanneeneutre = "10-01",
  save = FALSE,
  projet = NA_character_,
  format = ".png"
)
}
\arguments{
\item{data}{Data.frame issu de chronique.mesures (données fines)}

\item{titre}{Titre du graphique (vide par défaut)}

\item{origine_donnees}{Éventuelle source des données à afficher sur la figure (vide par défaut)}

\item{typemesure}{Défini le type de données et modifie les légendes en fonction. Ignoré si le champ chmes_typemesure est présent dans data}

\item{stations}{Dataframe contenant les stations, avec au minimum \code{chsta_coderhj} et \code{chsta_distancesource}}

\item{affichagevide}{Si \code{TRUE}, fait apparaître les années/stations ne contenant pas de résultats. Si \code{FALSE} (par défault), ne fait apparaître que les années/stations contenants des résultats}

\item{datedebutanneebiol}{Date de démarrage de l'année biologique : 10-01 (par défaut - 1er octobre), pour l'affichage sous forme de courbes}

\item{datedebutanneeneutre}{Date de démarrage de l'année neutre : 10-01 (par défaut - 1er octobre), pour l'affichage sous forme d'années neutres}

\item{save}{Si \code{FALSE} (par défault), n'enregistre pas les figures. Si \code{TRUE}, les enregistre.}

\item{projet}{Nom du projet}

\item{format}{Défini le format d'enregistrement (par défaut .png)}
}
\description{
Cette fonction permet de représenter les classes de valeurs de référence pour des chroniques de mesures (température, niveaux, etc.) sous forme de densités
}
\examples{
mesures_exemple \%>\% filter(chmes_coderhj == "HER14-8") \%>\% filter(chmes_validation == "Validé") \%>\% filter(chmes_typemesure == "Thermie") \%>\% chronique.figure.classesdensite()
mesures_exemple \%>\% filter(chmes_coderhj == "HER14-8") \%>\% filter(chmes_validation == "Validé") \%>\% filter(chmes_typemesure == "Thermie") \%>\% formatage.annee.biologique() \%>\% filter(chmes_anneebiol == 2021) mutate(mois = format(chmes_date, "\%m")) \%>\% filter(mois \%in\% c("06", "07", "08", "09")) \%>\% chronique.figure.classesdensite()
}
\keyword{chronique}
