% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.figure.cumul.R
\name{chronique.figure.cumul}
\alias{chronique.figure.cumul}
\title{Représentation de chroniques cumulées}
\usage{
chronique.figure.cumul(
  data = data,
  Titre = "",
  typemesure = c("Thermie", "Thermie barométrique", "Thermie piézométrique",
    "Oxygénation", "Hydrologie", "Pluviométrie"),
  duree = c("Relatif", "Complet"),
  datedebutanneebiol = "10-01",
  legendemax = 15,
  Ymin = -10,
  Ymax = NA,
  save = F,
  projet = NA_character_,
  etiquette = F,
  format = ".png"
)
}
\arguments{
\item{data}{Data.frame de données de mesures brutes, issues de chronique.mesures()}

\item{Titre}{Titre du graphique (vide par défaut)}

\item{typemesure}{Défini le type de données et modifie les légendes en fonction. Ignoré si le champ chmes_typemesure est présent dans data}

\item{duree}{Si \code{Relatif} (par défault), affichage uniquement de la période concernée. Si \code{Complet}, affichage de l'année complète.}

\item{datedebutanneebiol}{Date de démarrage de l'année biologique : 10-01 (par défaut - 1er octobre)}

\item{legendemax}{Nombre d'éléments maximum dans la légende (15 par défaut)}

\item{Ymin}{Valeur minimale de l'axe des Y (-1 par défaut)}

\item{Ymax}{Valeur maximale de l'axe des Y (aucune par défaut)}

\item{save}{Si \code{FALSE} (par défault), n'enregistre pas les figures. Si \code{TRUE}, les enregistre.}

\item{projet}{Nom du projet}

\item{etiquette}{Affichage des étiquettes de chroniques (\code{FALSE} (par défault))}

\item{format}{Défini le format d'enregistrement (par défaut .png)}
}
\description{
Cette fonction permet de représenter des cumuls de chroniques (température, débits, pluviométrie, etc.)
}
\examples{
chronique.figure.cumul(data)
MesuresTest \%>\% chronique.figure.cumul()
MesuresTest \%>\% chronique.agregation(datedebutanneebiol = "03-01") \%>\% purrr::pluck(2) \%>\% chronique.figure.cumul(datedebutanneebiol = "03-01")
}
\keyword{chronique}
