% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.figure.depassementscalendaires.R
\name{chronique.figure.depassementscalendaires}
\alias{chronique.figure.depassementscalendaires}
\title{Dépassements de valeurs seuils de chroniques}
\usage{
chronique.figure.depassementscalendaires(
  data = data,
  Titre = "",
  typemesure = c("Thermie", "Thermie barométrique", "Thermie piézométrique",
    "Barométrie", "Piézométrie", "Piézométrie brute", "Piézométrie compensée",
    "Piézométrie calée", "Piézométrie NGF", "Oxygénation", "Hydrologie",
    "Pluviométrie"),
  seuil_variable = c("VMedJ", "VMinJ", "VMoyJ", "VMaxJ", "VAmpliJ", "VAmpliSigneJ",
    "VarJ", "NMesuresJ", "SommeMoyJ"),
  seuil_valeur = 19,
  seuil_limite_exces = 12,
  etiquette_superieure = "Supérieur",
  couleur_superieure = "#FDAE61",
  etiquette_inferieure = "Inférieur",
  couleur_inferieure = "#2B83BA",
  affichagevide = FALSE,
  datedebutanneebiol = "10-01",
  save = F,
  projet = NA_character_,
  format = ".png"
)
}
\arguments{
\item{data}{Data.frame issu de chronique.agregation (données journalières)}

\item{Titre}{Titre du graphique (vide par défaut)}

\item{typemesure}{Défini le type de données et modifie les légendes en fonction. Ignoré si le champ chmes_typemesure est présent dans data}

\item{seuil_variable}{Type de variable journalière agrégée à utiliser : \code{VMedJ} (par défaut), \code{VMinJ}, \code{VMoyJ}, \code{VMaxJ}, \code{VAmpliJ}, \code{VAmpliSigneJ}, \code{VarJ}, \code{NMesuresJ} ou \code{SommeMoyJ}}

\item{seuil_valeur}{Valeur de seuil à tester (valeur de \code{19} par défaut). Le calcul réalise un ≥.}

\item{seuil_limite_exces}{Seuil en-deça duquel les dépassements sont testés par défaut, sinon ils sont testés par excès (valeur de \code{12} par défaut).}

\item{etiquette_superieure}{Intitulé à afficher en cas de dépassement par excès (\code{Supérieur} par défaut)}

\item{couleur_superieure}{Couleur à afficher en cas de dépassement par excès (\code{#FDAE61} - orange par défaut)}

\item{etiquette_inferieure}{Intitulé à afficher en cas de dépassement par défaut (\code{Inférieur} par défaut)}

\item{couleur_inferieure}{Couleur à afficher en cas de dépassement par défaut (\code{#2B83BA} - bleu par défaut)}

\item{affichagevide}{Si \code{TRUE}, fait apparaître les années ne contenant pas de résultats. Si \code{FALSE} (par défault), ne fait apparaître que les années contenants des résultats}

\item{datedebutanneebiol}{Date de démarrage de l'année biologique : 10-01 (par défaut - 1er octobre), pour l'affichage sous forme de courbes}

\item{save}{Si \code{FALSE} (par défault), n'enregistre pas les
figures. Si \code{TRUE}, les enregistre.}

\item{projet}{Nom du projet}

\item{format}{Défini le format d'enregistrement (par défaut .png)}
}
\description{
Cette fonction permet de représenter d'éventuels dépassements de valeurs de référence pour des chroniques de mesures (température, niveaux, etc.) sous forme de comparaison interannuelle
}
\examples{
data \%>\% chronique.agregation() \%>\% purrr::pluck(2) \%>\% chronique.figure.depassementscalendaires()
data \%>\% chronique.agregation(instantanne = F, mensuel = F, annuel = F, integral = F) \%>\% chronique.figure.depassementscalendaires()
MesuresCaleesAgregeesJours \%>\% chronique.figure.depassementscalendaires(typemesure = "Piézométrie NGF", seuil_valeur = 477.25, etiquette_superieure = "Hautes-eaux", etiquette_inferieure = "Basses-eaux")
}
\keyword{chronique}
