% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.figure.horizonplot.R
\name{chronique.figure.horizonplot}
\alias{chronique.figure.horizonplot}
\title{Vue sous forme d'horizon plot}
\usage{
chronique.figure.horizonplot(
  data = data,
  titre = "",
  origine_donnees = "",
  typemesure = c("Thermie", "Thermie barométrique", "Thermie piézométrique",
    "Barométrie", "Piézométrie", "Piézométrie brute", "Piézométrie compensée",
    "Piézométrie calée", "Piézométrie NGF", "Oxygénation", "Hydrologie",
    "Pluviométrie"),
  stations = NA_character_,
  affichagevide = FALSE,
  datedebutanneebiol = "10-01",
  datedebutanneeneutre = "10-01",
  save = FALSE,
  projet = NA_character_,
  format = ".png"
)
}
\arguments{
\item{data}{Data.frame issu de chronique.agregation (données journalières)}

\item{titre}{Titre du graphique (vide par défaut)}

\item{origine_donnees}{Éventuelle source des données à afficher sur la figure (vide par défaut)}

\item{typemesure}{Défini le type de données et modifie les légendes en fonction. Ignoré si le champ chmes_typemesure est présent dans data}

\item{stations}{Dataframe contenant les stations, avec au minimum \code{chsta_coderhj} et \code{chsta_distancesource}}

\item{affichagevide}{Si \code{TRUE}, fait apparaître les années/stations ne contenant pas de résultats. Si \code{FALSE} (par défault), ne fait apparaître que les années/stations contenants des résultats}

\item{datedebutanneebiol}{Date de démarrage de l'année biologique : 10-01 (par défaut - 1er octobre), pour l'affichage sous forme de courbes}

\item{datedebutanneeneutre}{Date de démarrage de l'année neutre : 10-01 (par défaut - 1er octobre), pour l'affichage sous forme d'années neutres}

\item{save}{Si \code{FALSE} (par défault), n'enregistre pas les figures. Si \code{TRUE}, les enregistre.}

\item{projet}{Nom du projet}

\item{format}{Défini le format d'enregistrement (par défaut .png)}
}
\description{
Cette fonction permet de représenter les classes de valeurs des chroniques de mesures (température, niveaux, etc.) sous forme d'un graphique d'horizons
}
\examples{
mesures_exemple \%>\% filter(grepl("DRO11-6", chmes_coderhj)) \%>\% filter(chmes_validation == "Validé") \%>\% filter(chmes_typemesure == "Thermie") \%>\% chronique.agregation(instantanne = FALSE, mensuel = FALSE, annuel = FALSE, integral = FALSE, complement = TRUE) \%>\% formatage.annee.biologique(datedebutanneebiol = "01-01") \%>\% formatage.annee.neutre() \%>\% chronique.figure.horizonplot()
}
\keyword{chronique}
