% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.figure.interannuelle.R
\name{chronique.figure.interannuelle}
\alias{chronique.figure.interannuelle}
\title{Comparaison interannuelles de chroniques}
\arguments{
\item{data}{Data.frame contenant a minima une colonne chmes_date et une colonne chmes_valeur}

\item{Titre}{Titre du graphique (vide par défaut)}

\item{typemesure}{Défini le type de données et modifie les légendes en fonction. Ignoré si le champ chmes_typemesure est présent dans data}

\item{Vmm30j}{Si \code{FALSE} (par défault), n'affiche pas les
Vmm30j.  Si \code{TRUE}, les affiche.}

\item{Vminmax}{Si \code{TRUE} (par défault), affiche pas les
valeurs journalières minimales et maximales.  Si \code{FALSE}, ne les affiche pas.}

\item{Ymin}{Valeur minimale de l'axe des Y (-1 par défaut)}

\item{Ymax}{Valeur maximale de l'axe des Y (aucune par défaut)}

\item{style}{En forme de boxplot (par défaut) ou de violon}

\item{save}{Si \code{FALSE} (par défault), n'enregistre pas les
figures.  Si \code{TRUE}, les enregistre.}

\item{projet}{Nom du projet}

\item{format}{Défini le format d'enregistrement (par défaut .png)}
}
\description{
Cette fonction permet de représenter des chroniques de mesures (température, niveaux, etc.) sous forme de comparaison interannuelle
}
\examples{
chronique.figure.interannuelle(data)
data \%>\% chronique.figure.interannuelle()
chronique.figure.interannuelle(data = tableaudonnee, Titre=nom, typemesure = "Barométrie", save=T, format=".png")
}
\keyword{chronique}
