% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.figure.longitudinale.R
\name{chronique.figure.longitudinale}
\alias{chronique.figure.longitudinale}
\title{Profils en long de chroniques}
\usage{
chronique.figure.longitudinale(
  data = data,
  save = F,
  projet = NA_character_,
  format = ".png"
)
}
\arguments{
\item{data}{Data.frame issu de la fonction chronique.resultats ou de la base de données Chroniques, avec jointure de chsta_milieu et de chsta_distancesource}

\item{save}{Si \code{FALSE} (par défault), n'enregistre pas les
figures. Si \code{TRUE}, les enregistre.}

\item{projet}{Nom du projet}

\item{format}{Défini le format d'enregistrement (par défaut .png)}
}
\description{
Cette fonction permet de représenter sous forme de profil longitudinal des données de chroniques
}
\examples{
Resultats \%>\% chronique.figure.longitudinale()
Resultats \%>\% chronique.resultats.filtrage() \%>\% chronique.figure.longitudinale()
Resultats \%>\% left_join(Stations \%>\% select(chsta_coderhj, chsta_milieu, chsta_distancesource, chsta_distancesource_confluencedrainprincipal), by = c("chres_coderhj" = "chsta_coderhj")) \%>\% chronique.resultats.filtrage() \%>\% chronique.figure.longitudinale()
}
\keyword{chronique}
