% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.figure.parametres.R
\name{chronique.figure.parametres}
\alias{chronique.figure.parametres}
\title{Paramétrisation des vues de chroniques}
\usage{
chronique.figure.parametres(
  data = data,
  typefigure = c("valeurs", "cumul", "vmm30j")
)
}
\arguments{
\item{data}{Data.frame issu de la fonction chronique.contexte()}

\item{typefigure}{Type de figure qui est à paramétriser : \code{valeurs} (par défault), \code{cumul} ou \code{vmm30j}}
}
\description{
Cette fonction permet de sélectionner automatiquement les paramètres de légende des représentations graphiques de chroniques à partir des données issues de chronique.contexte
}
\examples{
chronique.mesures("DRO6-8") \%>\% chronique.contexte() \%>\% chronique.figure.parametres()

parametres <- chronique.mesures("DRO6-8") \%>\% chronique.contexte() \%>\% chronique.figure.parametres()
unite <- parametres$unite
legendeY <- parametres$legendeY
legendeTitre <- parametres$legendeTitre
typemesureTitreSortie <- parametres$typemesureTitreSortie
}
\keyword{chronique}
