% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.figure.pluviometrie.R
\name{chronique.figure.pluviometrie}
\alias{chronique.figure.pluviometrie}
\title{Représentation de chroniques de pluviométrie}
\usage{
chronique.figure.pluviometrie(
  data = data,
  Titre = "",
  type = c("montant", "descendant"),
  datedebutanneebiol = "10-01",
  save = F,
  projet = NA_character_,
  format = ".png"
)
}
\arguments{
\item{data}{Data.frame de données de mesures brutes, issues de chronique.mesures()}

\item{Titre}{Titre du graphique (vide par défaut)}

\item{type}{Type de graphique : \code{montant} (par défaut), ou \code{descendant}}

\item{datedebutanneebiol}{Date de démarrage de l'année biologique : 10-01 (par défaut - 1er octobre)}

\item{save}{Si \code{FALSE} (par défault), n'enregistre pas les figures. Si \code{TRUE}, les enregistre.}

\item{projet}{Nom du projet}

\item{format}{Défini le format d'enregistrement (par défaut .png)}
}
\description{
Cette fonction permet de représenter des chroniques de pluviométrie
}
\examples{
data \%>\% chronique.figure.pluviometrie()
MesuresTest \%>\% chronique.figure.pluviometrie(datedebutanneebiol = "01-01")
MesuresTest \%>\% chronique.figure.pluviometrie(datedebutanneebiol = "01-01") + scale_x_date(limits = c(ymd("2023-01-01"), ymd("2023-12-31")), labels = scales::date_format("\%b", locale = "fr"), date_minor_breaks = "1 month")
}
\keyword{chronique}
