% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.figure.R
\name{chronique.figure}
\alias{chronique.figure}
\title{Représentation de chroniques}
\usage{
chronique.figure(thermie = tab, CodeRDT = "",
  legendeY = "Température (°C)", duree = "Complet", save = F,
  format = ".png")
}
\arguments{
\item{thermie}{Data.frame contenant a minima une colonne Date et une colonne Temp}

\item{CodeRDT}{Titre du graphique (vide par défaut)}

\item{legendeY}{Défini le texte de la légende de l'axe Y (Température (C))}

\item{duree}{Si \code{Complet} (par défault), affichage de l'année complète.  Si \code{Relatif}, affichage uniquement de la période concernée.}

\item{save}{Si \code{FALSE} (par défault), n'enregistre pas les
figures.  Si \code{TRUE}, les enregistre.}

\item{format}{Défini le format d'enregistrement (par défaut .png)}
}
\description{
Cette fonction permet de représenter des chroniques de mesures (température, niveaux, etc.)
}
\examples{
chronique.figure(thermie)
chronique.figure(thermie = mesdonnees, legendeY = "Température (°C)", duree = "Complet")
chronique.figure(thermie = tableaudonnee, CodeRDT=nom, legendeY = "Température (°C)", save=T, format=".png")
}
\keyword{temperature}

