% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.figureglobale.R
\name{chronique.figureglobale}
\alias{chronique.figureglobale}
\title{Représentation de chroniques}
\usage{
chronique.figureglobale(thermie = tab, st = "",
  legendeY = "Température (°C)", save = F, format = ".png")
}
\arguments{
\item{thermie}{Data.frame contenant a minima une colonne Date et une colonne Temp}

\item{st}{Titre du graphique (vide par défaut)}

\item{legendeY}{Défini le texte de la légende de l'axe Y (Température (C))}

\item{save}{Si \code{FALSE} (par défault), n'enregistre pas les
figures.  Si \code{TRUE}, les enregistre.}

\item{format}{Défini le format d'enregistrement (par défaut .png)}
}
\description{
Cette fonction permet de représenter des chroniques de mesures (température, niveaux, etc.)
}
\examples{
chronique.figureglobale(thermie)
chronique.figureglobale(thermie = mesdonnees, legendeY = "Température (°C)")
chronique.figureglobale(thermie = tableaudonnee, st=nom, legendeY = "Température (°C)", save=T, format=".png")
}
\keyword{temperature}

