% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.meteofrance.commande.R
\name{chronique.meteofrance.commande}
\alias{chronique.meteofrance.commande}
\title{Collecte d'une commande de données Météo-France}
\usage{
chronique.meteofrance.commande(
  id_commande = NA_character_,
  token = NA_character_,
  temporisation = 2
)
}
\arguments{
\item{id_commande}{Identifiant de commande, issu de la fonction \code{chronique.meteofrance.mesures}}

\item{token}{Token \code{API Key} créé dans son compte personnel du site \url{https://portail-api.meteofrance.fr}}

\item{temporisation}{Délai (en secondes) de temporisation après l'envoi de la commande, afin de réduire les erreurs 204 de l'API suite à une demande de téléchargement trop rapide}
}
\description{
Cette fonction permet de collecter les données Météo-France à partir d'un numéro de commande
}
\examples{
chronique.meteofrance.commande("2025003064730", token)
chronique.meteofrance.mesures("Quotidien", "39159002", "2025-01-01T00:00:00Z", "2025-04-01T00:00:00Z", token) \%>\% chronique.meteofrance.commande(token) \%>\% chronique.meteofrance.nettoyage()
}
