% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.meteofrance.nettoyage.R
\name{chronique.meteofrance.nettoyage}
\alias{chronique.meteofrance.nettoyage}
\title{Nettoyage de données Météo-France}
\usage{
chronique.meteofrance.nettoyage(
  data = NA_character_,
  format = c("Long", "Large")
)
}
\arguments{
\item{data}{Données brutes issues de \code{chronique.meteofrance.commande()}}

\item{format}{Format de sortie : \code{Long} par défaut, ou \code{Large}}
}
\description{
Cette fonction permet de nettoyer des données Météo-France collectées à partir de l'API
}
\examples{
chronique.meteofrance.mesures("Quotidien", "39159002", "2025-01-01T00:00:00Z", "2025-04-01T00:00:00Z", token) \%>\% chronique.meteofrance.commande(token) \%>\% chronique.meteofrance.nettoyage()
chronique.meteofrance.mesures("Quotidien", "39159002", "2025-01-01T00:00:00Z", "2025-04-01T00:00:00Z", token) \%>\% chronique.meteofrance.commande(token) \%>\% chronique.meteofrance.nettoyage(format = "Large")
}
