% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.meteofrance.stations.R
\name{chronique.meteofrance.stations}
\alias{chronique.meteofrance.stations}
\title{Liste des stations Météo-France}
\usage{
chronique.meteofrance.stations(
  type = c("6_minutes", "Horaire", "Quotidien"),
  departement_insee = NA_integer_,
  typemesure = c("Précipitations", "Température", "Humidité", "Vent", "Pression",
    "Rayonnement", "Insolation", "État de la mer", "ETP quotidienne"),
  stations_actives = TRUE,
  token = NA_character_
)
}
\arguments{
\item{type}{Fréquence de mesure recherchée : données à pas de temps (\code{Quotidien}) par défaut, \code{6_minutes}, \code{Horaire})}

\item{departement_insee}{Code Insee du département recherché}

\item{typemesure}{Type de mesure recherché : \code{Précipitations} par défaut, \code{Température}, \code{Humidité}, \code{Vent}, \code{Pression}, \code{Rayonnement}, \code{Insolation}, \code{État de la mer} et \code{ETP quotidienne}}

\item{stations_actives}{Filtrage des stations actives \code{TRUE} par défaut, \code{FALSE}}

\item{token}{Token \code{API Key} créé dans son compte personnel du site \url{https://portail-api.meteofrance.fr}}
}
\description{
Cette fonction permet de lister les stations Météo-France via l'API
}
\examples{
chronique.meteofrance.stations("Quotidien", 39, "Précipitations", token)
}
