% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.ouverture.R
\name{chronique.ouverture}
\alias{chronique.ouverture}
\title{Ouverture de fichiers de chroniques}
\arguments{
\item{localisation}{Localisation relative du fichier (à partir de /NAS-dataaimporter/)}
}
\description{
Cette fonction permet d'ouvrir de manière semi-automatisée des fichiers de chroniques
}
\examples{
chronique.ouverture("/NAS-dataaimporter/Chroniques/Saisie_JB/2018-12-17_Suivi_Ain_amont_FJPPMA_été_2018/10880567.txt")
}
\keyword{chronique}
