% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.recalage.valeur.R
\name{chronique.recalage.valeur}
\alias{chronique.recalage.valeur}
\title{Recalage des valeurs de données brutes de chronique}
\usage{
chronique.recalage.valeur(data, decalage = NA_real_)
}
\arguments{
\item{data}{Chronique à modifier, avec un champ chmes_date, un champ chmes_heure et un champ chmes_valeur}

\item{decalage}{Valeur de décalage. Si \code{vide} (par défaut), utilise la première valeur de la série pour atteindre 0}
}
\description{
Permet de recaler les valeurs de données brutes de chronique. ATTENTION : ne gère pas les ID
}
\examples{
chronique.recalage.valeur(data)
chronique.recalage.valeur(data, decalage = 258.58)
chronique.recalage.valeur(decalage = -57)
}
\keyword{chronique}
