% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.regroupement.R
\name{chronique.regroupement}
\alias{chronique.regroupement}
\title{Regroupement de données temporelles}
\description{
Permet de mettre bout à bout des données temporelles
}
\examples{
synthese <- chronique.regroupement()
}
\keyword{donnees}
