% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.resultats.filtrage.R
\name{chronique.resultats.filtrage}
\alias{chronique.resultats.filtrage}
\title{Filtrage des résultats de chroniques}
\usage{
chronique.resultats.filtrage(
  data = data,
  anneevmm = TRUE,
  datefperiode = "07-15",
  nbj = 75
)
}
\arguments{
\item{data}{Data.frame issu de la fonction chronique.resultats ou de la base de données Chroniques}

\item{anneevmm}{Si \code{TRUE} (par défault), ne conserve que les résultats dont l'année de fin de vmm30j est égale à l'année biologique du résultat}

\item{datefperiode}{Ne conserver que les résultats dont la date de fin de période est postérieure à la date choisie ("07-15", soit le 15 juillet par défault)}

\item{nbj}{Nombre de journées minimales pour chaque résultat à conserver (75 par défault).}
}
\description{
Cette fonction permet de filtrer les résultats de chroniques
}
\examples{
Resultats \%>\% chronique.resultats.filtrage()
Resultats \%>\% chronique.resultats.filtrage(datefperiode = "08-01")
}
\keyword{chronique}
