% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.resultats.periode.R
\name{chronique.resultats.periode}
\alias{chronique.resultats.periode}
\title{Périodes des résultats de chroniques}
\usage{
chronique.resultats.periode(
  data = data,
  typeperiode = c("Detail", "Enveloppe")
)
}
\arguments{
\item{data}{Data.frame issu de la fonction chronique.resultats ou de la base de données Chroniques}

\item{typeperiode}{Type de période que l'on souhaite calculer : période totale (\code{Enveloppe} : 2014-2018), ou période fractionnée (\code{Detail} : 2014-2015;2017;2019-2020, par défaut)}
}
\description{
Cette fonction permet de calculer les périodes couvertes par les résultats de chroniques (champs PeriodeTotale)
}
\examples{
Resultats \%>\% chronique.resultats.periode()
chronique.resultats("2017","Année biologique") \%>\% chronique.resultats.periode()
}
\keyword{chronique}
