% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroniques.resultats.R
\name{chronique.resultats}
\alias{chronique.resultats}
\title{Listage des résultats de chroniques}
\arguments{
\item{Nom}{Nom recherché}

\item{Recherche}{Type de données recherchées}
}
\description{
Cette fonction permet de lister les résultats de la BDD Chroniques
}
\examples{
chronique.capteurs("HER0-6","Station")
chronique.capteurs("2017","Année biologique")
}
\keyword{capteurs}
