% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.stations.R
\name{chronique.stations}
\alias{chronique.stations}
\title{Listage des stations de chroniques}
\arguments{
\item{Territoire}{Territoire concerné. Unique ou sous forme de vecteur (c("Villerserine", "Villers-Robert") par exemple)}

\item{Echelle}{Échelle spatiale de la synthèse (commune, canton, communauté de communes, département, région, contexte de PDPG, Hydro-écorégion, entité GEMAPI, Maître d'ouvrage, Milieu, Bassin versant, Sous-bassin versant, Polygone autre)}

\item{Sortie}{Forme du dataframe de sortie - \code{Simple} (par défault), \code{Propre} (format diffusable, avec stations) ou \code{Complet} (tous les champs)}
}
\description{
Cette fonction permet de lister les stations de la BDD Chroniques
}
\examples{
chronique.stations("CD39","MO")
chronique.stations("Suran","Milieu")
chronique.stations("Ain","Bassin")
chronique.stations("Serpentine","Sous-bassin")
chronique.stations("Lons-le-Saunier","Commune")
chronique.stations("39","Département")
}
\keyword{stations}
