% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.suivi.R
\name{chronique.suivi}
\alias{chronique.suivi}
\title{Extraction suivi terrain chroniques}
\arguments{
\item{Type}{Type de donnée de suivi}
}
\description{
Cette fonction permet d'extraire des données de suivi de terrain
}
\examples{
chronique.suivi("CG39", Type = "MO")
chronique.suivi("JB-Stéphane", Type = "Opérateur")
chronique.suivi("DRO14-2", Type = "Station")
chronique.suivi("2015-10-23", Type = "Date")
chronique.suivi("9759803", Type = "Sonde")
}
