% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.suppression.R
\name{chronique.suppression}
\alias{chronique.suppression}
\title{Suppression de données brutes de chronique}
\usage{
chronique.suppression(
  data,
  cCodeRDT = "BONbaro",
  cCapteur = "P0352",
  cDate = "2014-12-27",
  cHeure = "19:00:00",
  cTypeMesure = "Thermie",
  cValeur = 15,
  cValidation = "Validé"
)
}
\arguments{
\item{data}{Chronique à modifier}

\item{cCodeRDT}{chmes_coderhj à modifier}

\item{cCapteur}{chmes_capteur à modifier}

\item{cDate}{chmes_date à modifier}

\item{cHeure}{chmes_heure à modifier}

\item{cTypeMesure}{\code{Thermie} (par défault)}

\item{cValeur}{Valeur à supprimer}
}
\description{
Permet de supprimer des données brutes de chronique présentant des valeurs aberrantes. ATTENTION : ne gère pas les ID
}
\examples{
chronique.suppression(data, cCodeRDT = "BONbaro", cCapteur = "P0352", cDate = "2014-12-27", cHeure = "19:00:00", cTypeMesure = "Thermie", cValeur = 15, cValidation = "Validé")
}
\keyword{chronique}
