% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.tendance.R
\name{chronique.tendance}
\alias{chronique.tendance}
\title{Tendance d'évolution d'une chroniques}
\arguments{
\item{data}{Jeu de données}

\item{N}{}
}
\description{
Cette fonction permet d'analyser les tendances d'évolution de chroniques de mesures (température, niveaux, etc.)
}
\examples{
chronique.tendance(data)
chronique.tendance(data, N=5)
}
\keyword{chronique}
