% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.variables.renommage.R
\name{chronique.variables.renommage}
\alias{chronique.variables.renommage}
\title{Renommage des variables de chroniques}
\usage{
chronique.variables.renommage(
  data = data,
  formatentree = c("Tous", "chmes", "chmesgr", "chsta", "chsvi", "chres"),
  formatsortie = c("chmes", "chmesgr", "chsta", "chsvi", "chres", "param")
)
}
\arguments{
\item{data}{Data.frame contenant des variables à renommer}

\item{formatentree}{Type de format de noms de champs en entrée : \code{Tous} (par défaut), ou bien uniquement pour convertir certains : \code{chmes}, \code{chmesgr}, \code{chsta}, \code{chsvi}, \code{chres}}

\item{formatsortie}{Type de format de noms de champs en sortie : \code{chmes} (par défaut), \code{chmesgr}, \code{chsta}, \code{chsvi}, \code{chres}, \code{param} (pour utilisation générique)}
}
\description{
Cette fonction permet de renommer des variables de traitement de données de chroniques
}
\examples{
Resultats \%>\% chronique.variables.renommage()
Resultats \%>\% chronique.variables.renommage(formatsortie = "chmesgr")
}
\keyword{chronique}
