% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.vue.dynamique.R
\name{chronique.vue.dynamique}
\alias{chronique.vue.dynamique}
\title{Affichage dynamique d'une chronique}
\usage{
chronique.vue.dynamique(data = data)
}
\arguments{
\item{data}{Jeu de données à représenter}
}
\description{
Cette fonction permet de représenter de manière dynamique une chronique
}
\examples{
chronique.vue.dynamique(data)
}
\keyword{chronique}
