% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.fonctions.parametres.R
\name{dev.fonctions.parametres}
\alias{dev.fonctions.parametres}
\title{Extraire les valeurs par défaut des paramètres d'une fonction}
\usage{
dev.fonctions.parametres(fonction)
}
\arguments{
\item{fonction}{Fonction sur laquelle travailler}
}
\description{
Cette fonction permet d'extraire les valeurs par défaut des paramètres d'une fonction
}
\examples{
dev.fonctions.parametres(chronique.DCE)
}
\keyword{Développement}
