% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.comparaison.champs.R
\name{df.comparaison.champs}
\alias{df.comparaison.champs}
\title{Comparaison de noms de champs}
\usage{
df.comparaison.champs(data_1, data_2)
}
\arguments{
\item{data_1}{Jeu de données 1 à comparer}

\item{data_2}{Jeu de données 2 à comparer}
}
\description{
Cette fonction permet de comparer les champs présents dans deux dataframes (noms et datatypes)
}
\examples{
df.comparaison.champs(data_1, data_2)
data_1 \%>\% df.comparaison.champs(data_2) \%>\% view()
data_1 \%>\% df.comparaison.champs(data_2) \%>\% print(n = Inf)
}
\keyword{Données}
