% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/division.annuelle.R
\name{division.annuelle}
\alias{division.annuelle}
\title{Division en séries annuelles}
\description{
Permet de diviser une longue chronique temporelle en séries annuelles à partir du 1 octobre de l'année N-1 jusqu'au 30 septembre de l'année N
}
\examples{
division.annuelle()
}
\keyword{donnees}
