% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donnees.aberrantes.R
\name{donnees.aberrantes}
\alias{donnees.aberrantes}
\title{Détection des données aberrantes}
\arguments{
\item{data}{Jeu de données à traiter}
}
\description{
Cette fonction est une aide à la détection de données aberrantes dans une chronique
}
\examples{
donnees.aberrantes(data)
}
\keyword{Données}
