% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosystemes.affluents.R
\name{ecosystemes.affluents}
\alias{ecosystemes.affluents}
\title{Listage des affluents}
\usage{
ecosystemes.affluents(ecosysteme = "Ain")
}
\arguments{
\item{Nom}{du cours d'eau principal}
}
\description{
Cette fonction permet de lister les affluents d'un cours d'eau donné
}
\examples{
ecosystemes.affluents("Ain")
}
\keyword{écosystèmes}

