% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrage.estival.R
\name{filtrage.estival}
\alias{filtrage.estival}
\title{Filtrage des périodes estivales uniquement}
\usage{
filtrage.estival(data = data)
}
\arguments{
\item{data}{Jeu de données contenant une colonne \code{_date} au format Date}
}
\description{
Permet de ne conserver que les périodes estivales (mois 06, 07, 08 et 09)
}
\examples{
filtrage.estival(data)
mesures_annuelles \%>\% filtrage.estival()
}
\keyword{donnees}
