% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.abreviation.R
\name{formatage.abreviation}
\alias{formatage.abreviation}
\title{Remplacement d'abréviations}
\description{
Cette fonction permet d'exporter la définition des acronymes
}
\examples{
formatage.abreviation()

NomsColonnes <- colnames(Operations)
acronymes <- formatage.abreviation()
Operations <-
  Operations \%>\%
  left_join(acronymes, by = c(Station = "Acronyme")) \%>\%
  select(-Station) \%>\%
  rename(Station = Definition)
Operations <-
  Operations \%>\%
  select(match(NomsColonnes,names(.)))
}
