% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.annee.biologique.R
\name{formatage.annee.biologique}
\alias{formatage.annee.biologique}
\title{Reformatage de dates}
\usage{
formatage.annee.biologique(data)
}
\arguments{
\item{data}{Jeu de données contenant une colonne Date au format Date}
}
\description{
Reformate les dates par lot de fichiers (répertoire /Regroupement/Bruts vers /Regroupement/) en format "2014-10-10"
}
\examples{
formatage.annee.biologique(data)
}
\keyword{donnees}

