% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.annee.neutre.R
\name{formatage.annee.neutre}
\alias{formatage.annee.neutre}
\title{Reformatage de dates en année neutres}
\usage{
formatage.annee.neutre(data = data, datedebutanneeneutre = "10-01")
}
\arguments{
\item{data}{Jeu de données contenant une colonne chmes_date ou chsvi_date au format Date}

\item{datedebutanneeneutre}{Date de démarrage de l'année neutre : 10-01 (par défaut - 1er octobre)}
}
\description{
Reformate les dates selon une année neutre/arbitraire
}
\examples{
formatage.annee.neutre(data)
}
\keyword{donnees}
