% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.date.R
\name{formatage.date}
\alias{formatage.date}
\title{Reformatage de dates}
\usage{
formatage.date(style = "dmy")
}
\arguments{
\item{style}{Indique le style en entrée (dmy par défault)}
}
\description{
Reformate les dates par lot de fichiers (répertoire /Regroupement/Bruts vers /Regroupement/) en format "2014-10-10"
}
\examples{
formatage.date()
formatage.date(style = "ydm")
}
\keyword{donnees}
