% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.ecosysteme.R
\name{formatage.ecosysteme}
\alias{formatage.ecosysteme}
\title{Mise en acronymes des écosystèmes}
\usage{
formatage.ecosysteme(acronymes = data$Station)
}
\description{
Cette fonction permet de mettre en acronymes des noms d'écosystèmes
}
\examples{
formatage.ecosysteme(data$Station)
PC$CodeRDT <- formatage.ecosysteme(PC$CodeRDT)
}
