% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.heure.excel.R
\name{formatage.heure.excel}
\alias{formatage.heure.excel}
\title{Reformatage des heures issues d'excel}
\usage{
formatage.heure.excel(data = NA)
}
\arguments{
\item{data}{Heure à corriger (\code{0.57} par exemple)}
}
\description{
Cette fonction permet de mettre de reformater les heures d'excel importées au format numérique
}
\examples{
formatage.heure.excel(0.57)
}
