% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.heure.simple.R
\name{formatage.heure.simple}
\alias{formatage.heure.simple}
\title{Reformatage d'une colonne heure avec seulement les heures}
\usage{
formatage.heure.simple(data)
}
\arguments{
\item{data}{Jeu de données contenant une colonne \verb{_heure}}
}
\description{
Reformate une colonne heure avec seulement l'heure plein
}
\examples{
formatage.heure.simple(data)
}
\keyword{donnees}
