% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.liste.R
\name{formatage.liste}
\alias{formatage.liste}
\title{Reformatage sous forme d'une liste propre de valeurs}
\usage{
formatage.liste(data, colonne = NA_character_)
}
\arguments{
\item{data}{Jeu de données à traiter}

\item{colonne}{Colonne à traiter}
}
\description{
Reformate les valeurs d'une colonne sous la forme d'une liste propre
}
\examples{
data \%>\% formatage.liste(colonne)
stations_exemple \%>\% formatage.liste("chsta_mo")
}
\keyword{donnees}
