% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.mois.R
\name{formatage.mois}
\alias{formatage.mois}
\title{Reformatage de dates en mois}
\usage{
formatage.mois(data = data)
}
\arguments{
\item{data}{Jeu de données contenant une colonne \code{_date} au format Date}
}
\description{
Reformate les dates en mois dans une colonne \code{mois}
}
\examples{
formatage.mois(data)
mesures_annuelles \%>\% formatage.mois() \%>\% filter(mois \%in\% c("6", "7", "8", "9"))
}
\keyword{donnees}
