% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.time.R
\name{formatage.time}
\alias{formatage.time}
\title{Reformatage de dates et heures en une colonne time}
\usage{
formatage.time(data)
}
\arguments{
\item{data}{Jeu de données contenant une colonne \code{chmes_date}, \code{chmesgr_date}, \code{chsvi_date}, \code{chres_date}, \code{pcmes_date} ou \code{pcsvi_date} et une colonne \code{chmes_heure}, \code{chmesgr_heure}, \code{chsvi_heure}, \code{chres_heure}, \code{pcmes_heure} ou \code{pcsvi_heure}.}
}
\description{
Reformate les dates et les heures en une nouvelle colonne time
}
\examples{
formatage.time(data)
}
\keyword{donnees}
