% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.variables.renommage.R
\name{formatage.variables.renommage}
\alias{formatage.variables.renommage}
\title{Renommage des variables de chroniques et physico-chimie}
\usage{
formatage.variables.renommage(data = data)
}
\arguments{
\item{data}{Data.frame contenant des variables à renommer}
}
\description{
Cette fonction permet de renommer des variables de traitement de données de chroniques et de physico-chimie
}
\examples{
Resultats \%>\% formatage.variables.renommage()
}
\keyword{chronique}
