% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrobio.hubeau.R
\name{hydrobio.hubeau}
\alias{hydrobio.hubeau}
\title{Collecte de données de hydrobiologiques depuis Hub'eau}
\usage{
hydrobio.hubeau(
  departementinsee = NA_integer_,
  indicesandre = NA_integer_,
  codesie = NA_integer_,
  bbox = NA
)
}
\arguments{
\item{departementinsee}{Code INSEE du département (\code{39} par défaut)}

\item{indicesandre}{Code SANDRE de l'indice concerné : \code{1000} pour IBGN, \code{2928} pour IBMR, \code{5856} pour IBD, \code{7036} pour IPR, \code{7939} pour IPLAC}

\item{codesie}{Code SANDRE de la station concernée}

\item{bbox}{Zone géographique concernée, au format bbox généré  par sf, en WGS84/EPSG 4326}
}
\description{
Cette fonction permet de collecter de données d'hydrobiologie depuis Hub'eau
}
\examples{
hydrobio.hubeau(39, 7939)
hydrobio.hubeau(indicesandre = 7939)
hydrobio.hubeau(indicesandre = 7939) \%>\% filter(grepl("rousses", tolower(libelle_station_hydrobio)))
}
