% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrologie.enquetedeau.R
\name{hydrologie.enquetedeau}
\alias{hydrologie.enquetedeau}
\title{Collecte de données En quête d'eau}
\usage{
hydrologie.enquetedeau(
  code_departement = NA_integer_,
  annee = NA_integer_,
  formatage = TRUE
)
}
\arguments{
\item{code_departement}{Code INSEE du département (\code{vide} par défaut)}

\item{annee}{Année recherchée (\code{vide} par défaut)}

\item{formatage}{Formatage des champs au format interne FDPPMA39 (\code{TRUE} par défaut)}
}
\description{
Cette fonction permet de collecter des données d'observations En quête d'eau
}
\examples{
hydrologie.enquetedeau(39, 2026)
hydrologie.enquetedeau(39, 2026, formatage = FALSE)
}
