% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrologie.hubeau.onde.R
\name{hydrologie.hubeau.onde}
\alias{hydrologie.hubeau.onde}
\title{Collecte de données ONDE depuis Hub'eau}
\usage{
hydrologie.hubeau.onde(code_departement = NA_integer_, formatage = TRUE)
}
\arguments{
\item{code_departement}{Code INSEE du département (\code{vide} par défaut)}

\item{formatage}{Formatage des champs au format interne FDPPMA39 (\code{TRUE} par défaut)}
}
\description{
Cette fonction permet de collecter des données d'observations ONDE depuis Hub'eau
}
\examples{
hydrologie.hubeau.onde(39)
hydrologie.hubeau.onde(39, formatage = FALSE)
}
