% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrologie.hubeau.R
\name{hydrologie.hubeau}
\alias{hydrologie.hubeau}
\title{Collecte de données de hydrologiques depuis Hub'eau}
\usage{
hydrologie.hubeau(
  type = c("qmnj", "qmm", "hixm", "hixnj", "qinm", "qinnj", "qixm", "qixnj", "tr"),
  codesie = NA_character_,
  date_debut_obs = NA_character_,
  date_fin_obs = NA_character_,
  bbox = NA
)
}
\arguments{
\item{type}{Type de mesure recherche : débits moyens journaliers QmnJ (\code{qmnj}) par défaut, débits moyens mensuels (QmM) (\code{qmm}), Hauteur instantanée maximale mensuelle (HIXM) (\code{hixm}), Hauteur instantanée maximale journalière (HIXnJ) (\code{hixnj}), Débit instantané minimal mensuel (QINM) (\code{qinm}), Débit instantané minimal journalier (QINnJ) (\code{qinnj}), Débit instantané maximal mensuel (QixM) (\code{qixm}), Débit instantané maximal journalier (QIXnJ) (\code{qixnj}), temps réel (\code{tr})}

\item{codesie}{Code SIE de la station concernée}

\item{date_debut_obs}{Date de début des mesures à collecter (exprimée en TU), les formats de date (ISO 8601) supportés : \code{yyyy-MM-dd}, \code{yyyy-MM-dd'T'HH:mm:ss}, \code{yyyy-MM-dd'T'HH:mm:ssXXX}, exemples : \code{2018-12-01}, \code{2018-12-11T00:00:01}, \code{2018-12-11T00:00:01Z}}

\item{date_fin_obs}{Date de fin des mesures à collecter (exprimée en TU), les formats de date (ISO 8601) supportés : \code{yyyy-MM-dd}, \code{yyyy-MM-dd'T'HH:mm:ss}, \code{yyyy-MM-dd'T'HH:mm:ssXXX}, exemples : \code{2018-12-01}, \code{2018-12-11T00:00:01}, \code{2018-12-11T00:00:01Z}}

\item{bbox}{Zone géographique concernée, au format bbox généré  par sf, en WGS84/EPSG 4326}
}
\description{
Cette fonction permet de collecter de données d'hydrologie depuis Hub'eau
}
\examples{
hydrologie.hubeau("qmnj", "V244402001", "2025-04-01")
hydrologie.hubeau("qmnj", "V231000201", "2025-04-01", "2025-04-10")
hydrologie.hubeau("tr", "V231000201", today() - days(1)) \%>\% view()
}
