% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrologie.hubeau.stations.R
\name{hydrologie.hubeau.stations}
\alias{hydrologie.hubeau.stations}
\title{Collecte des stations hydrologiques depuis Hub'eau}
\usage{
hydrologie.hubeau.stations(
  departement_insee = NA_character_,
  commune_insee = NA_character_,
  region_insee = NA_character_,
  bbox = NA
)
}
\arguments{
\item{departement_insee}{Code INSEE du département concerné}

\item{commune_insee}{Code INSEE de la commune concernée}

\item{region_insee}{Code INSEE de la région concernée}

\item{bbox}{Zone géographique concernée, au format bbox généré par sf, en WGS84/EPSG 4326}
}
\description{
Cette fonction permet de collecter les stations de suivi hydrologique depuis Hub'eau
}
\examples{
hydrologie.hubeau.stations(39)
hydrologie.hubeau.stations(departement_insee = 25)
hydrologie.hubeau.stations(commune_insee = 39435)
hydrologie.hubeau.stations(region_insee = 27)
}
