% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrovu.authentification.R
\name{hydrovu.authentification}
\alias{hydrovu.authentification}
\title{Authentification à l'API HydroVu}
\usage{
hydrovu.authentification(id_user = NA_character_, id_key = NA_character_)
}
\arguments{
\item{id_user}{Identifiant de l'utilisateur autorisé sur l'API}

\item{id_key}{Clé de l'utilisateur autorisé sur l'API}
}
\description{
Cette fonction permet de s'identifier à l'API HydroVu
}
\examples{
hydrovu.authentification(id_user, id_key)
token <- hydrovu.authentification(id_user, id_key)
hydrovu.authentification("jbf", keyring::key_get("hydrovu", "jbf"))
"jbf" \%>\% hydrovu.authentification(., keyring::key_get("hydrovu", .))
token <- "jbf" \%>\% hydrovu.authentification(., keyring::key_get("hydrovu", .))
}
\keyword{chronique}
