% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrovu.donnees.R
\name{hydrovu.donnees}
\alias{hydrovu.donnees}
\title{Collecte des données de l'API HydroVu}
\usage{
hydrovu.donnees(
  token = NA_character_,
  id = NA_character_,
  time_start = NA,
  time_end = NA
)
}
\arguments{
\item{token}{Identifiant de l'utilisateur autorisé sur l'API}

\item{id}{Id du site recherché sur l'API}

\item{time_start}{Date et heure de début de la chronique recherchée (format epoch \code{1683885600} ou ISO 8601 \code{2023-03-12 12:00:00})}

\item{time_end}{Date et heure de fin de la chronique recherchée (format epoch \code{1683885600} ou ISO 8601 \code{2023-03-12 12:00:00})}
}
\description{
Cette fonction permet de collecter les données disponibles sur l'API HydroVu
}
\examples{
hydrovu.donnees(token, "4722100874248192")
"4722100874248192" \%>\% hydrovu.donnees(token, .)
hydrovu.authentification(id_user, id_key) \%>\% hydrovu.donnees("4722100874248192")
hydrovu.authentification(id_user, id_key) \%>\% hydrovu.donnees("4722100874248192", 1683885600, 1683892800)
hydrovu.donnees("4722100874248192", 1683885600, 1683892800) \%>\% # Ok
hydrovu.donnees("4722100874248192", "2023-05-12 10:00:00", 1683892800) \%>\% 
hydrovu.donnees("4722100874248192", 1683880600, "2023-05-12 12:00:00") \%>\%
hydrovu.donnees("4722100874248192", "2023-05-12 10:00:00", "2023-05-12 12:00:00")
}
\keyword{chronique}
