% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrovu.extraction.R
\name{hydrovu.extraction}
\alias{hydrovu.extraction}
\title{Extraction en masse des données de l'API HydroVu}
\usage{
hydrovu.extraction(
  token = NA_character_,
  id = NA_character_,
  time_start = NA,
  time_end = NA
)
}
\arguments{
\item{token}{Identifiant de l'utilisateur autorisé sur l'API}

\item{id}{Id du site recherché sur l'API}

\item{time_start}{Date et heure de début de la chronique recherchée (format epoch \code{1683885600} ou ISO 8601 \code{2023-03-12 12:00:00})}

\item{time_end}{Date et heure de fin de la chronique recherchée (format epoch \code{1683885600} ou ISO 8601 \code{2023-03-12 12:00:00})}
}
\description{
Cette fonction permet d'extraire en masse les données disponibles l'API HydroVu
}
\examples{
tibble(chmes_coderhj = "4722100874248192", time = seq(ymd_hms('2022-01-01 00:00:00'), ymd_hms('2022-10-31 00:00:00'), by='2 hour')) \%>\% head(25) \%>\% group_split(chmes_coderhj, time) \%>\% map(~ hydrovu.extraction(token, .$chmes_coderhj, .$time)) \%>\% list_rbind() 
}
\keyword{chronique}
