% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrovu.refactoring.R
\name{hydrovu.refactoring}
\alias{hydrovu.refactoring}
\title{Refactoring des données de l'API HydroVu}
\usage{
hydrovu.refactoring(data)
}
\arguments{
\item{data}{Données collectées sur l'API via \code{hydrovu.donnees}}
}
\description{
Cette fonction permet de reformater les données collectées sur l'API HydroVu
}
\examples{
hydrovu.authentification(id_user, id_key) \%>\% hydrovu.donnees("4722100874248192") \%>\% hydrovu.refactoring()
}
\keyword{chronique}
