% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrovu.sites.R
\name{hydrovu.sites}
\alias{hydrovu.sites}
\title{Collecte des sites de l'API HydroVu}
\usage{
hydrovu.sites(token = NA_character_)
}
\arguments{
\item{token}{Identifiant de l'utilisateur autorisé sur l'API}
}
\description{
Cette fonction permet de collecter les sites disponibles sur l'API HydroVu
}
\examples{
hydrovu.sites(token)
hydrovu.authentification(id_user, id_key) \%>\% hydrovu.sites()
}
\keyword{chronique}
