% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.operations.R
\name{MI.operations}
\alias{MI.operations}
\title{Listing des opérations (MI) pour une station}
\usage{
MI.operations(listeStations = data.frame(Nom = character(0)),
  CodeOperation = F, Sortie = c("Simple", "Propre", "Complet"))
}
\arguments{
\item{listeStations}{Dataframe contenant un colonne "Nom" avec le code de la station (RHJ)}

\item{CodeOperation}{Affichage du CodeOperation - \code{FALSE} (par défault)}

\item{Sortie}{Forme du dataframe de sortie - \code{Simple} (par défault), \code{Propre} (format diffusable, avec stations) ou \code{Complet} (tous les champs)}
}
\description{
Cette fonction permet de lister les différentes opérations pour une liste donnée de stations
}
\examples{
MI.operations()
MI.operations(CodeOperation = TRUE)
MI.operations(data.frame(Nom = "BCH5-7"))
MI.operations(data.frame(Nom = "BCH5-7"), CodeOperation = TRUE)
MI.operations(listeStations, Sortie = Propre)
MI.operations(data.frame(Nom = "BCH5-7"), Sortie = Complet)
}
\keyword{poissons}
