% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.SANDRE.R
\name{MI.SANDRE}
\alias{MI.SANDRE}
\title{Ajout des taxons et codes SANDRE}
\usage{
MI.SANDRE(data)
}
\description{
Ajout des codes SANDRE et du nom du taxon de rang inférieur (espèce -> genre -> etc.) à partir d'un tableau issu de MI.systematique
}
\examples{
MI.SANDRE(data)
}
\keyword{donnees}
