% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.systematique.R
\name{MI.systematique}
\alias{MI.systematique}
\title{Complément des données de MI}
\arguments{
\item{data}{Jeu de données à compléter}
}
\description{
Cette fonction permet de compléter les données de systématique des MI
}
\examples{
MI.systematique(data)
}
\keyword{MI}
