% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/now.ymd_hm.R
\name{now.ymd_hm}
\alias{now.ymd_hm}
\title{Affichage de la date/heure actuelle}
\usage{
now.ymd_hm()
}
\description{
Cette fonction permet d'afficher la date/heure actuelle au format \code{ymd_hm}
}
\examples{
now.ymd_hm()
}
\keyword{chronique}
