% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/now.ymd_hms.R
\name{now.ymd_hms}
\alias{now.ymd_hms}
\title{Affichage de la date/heure actuelle}
\usage{
now.ymd_hms()
}
\description{
Cette fonction permet d'afficher la date/heure actuelle au format \code{ymd_hms}
}
\examples{
now.ymd_hm()
}
\keyword{chronique}
