% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.agregation.R
\name{PC.agregation}
\alias{PC.agregation}
\title{Agrégation de paramètres physio-chimiques}
\arguments{
\item{PC}{Jeu de données}
}
\description{
Cette fonction permet de regrouper certains paramètres physico-chimiques (PCB 7, HAP 14, pesticides totaux, etc.)
}
\examples{
PC.agregation(PC)
}
