% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.classes.R
\name{PC.classes}
\alias{PC.classes}
\title{Attribution de classes de qualité}
\usage{
PC.classes(PC, Referentiel = c("NV", "SEQ-EAU", "Quebec"))
}
\arguments{
\item{PC}{Jeu de données}

\item{Referentiel}{Référentiel de qualité (NV par défaut)}
}
\description{
Cette fonction permet d'attribuer des classes de qualité à des valeurs physico-chimiques
}
\examples{
PC.classes(PC)
PC.classes(PC,Referentiel="NV")
PC.classes(PC,Referentiel="SEQ-EAU")
PC.classes(PC,Referentiel="Quebec")
}
