% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.concentrationO2.R
\name{PC.concentrationO2}
\alias{PC.concentrationO2}
\title{Saturation en oxygène}
\usage{
PC.concentrationO2(data)
}
\arguments{
\item{data}{Jeu de données (contenant une colonne Thermie et une colonne Saturation)}
}
\description{
Cette fonction permet de calculer des concentrations en oxygène à partir des chroniques de mesures de saturation et de température de l'eau
}
\examples{
PC.concentrationO2(data)
dataaimporter \%>\% PC.concentrationO2()
# Calcul à partir de https://mdm.sandre.eaufrance.fr/node/414781
}
\keyword{Physico-chimie}
