% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.conductivite.25degres.R
\name{PC.conductivite.25degres}
\alias{PC.conductivite.25degres}
\title{Conductivité corrigée à 25°C}
\usage{
PC.conductivite.25degres(data)
}
\arguments{
\item{data}{Jeu de données (contenant une colonne Thermie et une colonne Conductivité)}
}
\description{
Cette fonction permet de corriger (correction de température non linéaire) des conductivité vers des conductivités corrigées à 25°C au sens des normes ISO/DIN 7888 ou AFNOR NF en 2788
}
\examples{
PC.conductivite.25degres(data)
dataaimporter \%>\% PC.conductivite.25degres()
}
\keyword{Physico-chimie}
