% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.contexte.R
\name{PC.contexte}
\alias{PC.contexte}
\title{Contexte de données de physico-chimie}
\usage{
PC.contexte(data)
}
\arguments{
\item{data}{Data.frame issu des fonctions fd_production.physicochimie_mesures}
}
\description{
Cette fonction permet d'établir le contexte de données de physico-chimie (nombre de stations, de types de mesures, etc.)
}
\examples{
Contexte <- PC.contexte(Mesures)
}
\keyword{physico-chimie}
