% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.hubeau.operation.R
\name{PC.hubeau.operation}
\alias{PC.hubeau.operation}
\title{Collecte des opérations de physico-chimie depuis Hub'eau}
\usage{
PC.hubeau.operation(
  codesie = NA_character_,
  codeoperation = NA_character_,
  communeinsee = NA_character_,
  bbox = NA,
  date_start = NA_character_,
  date_end = NA_character_
)
}
\arguments{
\item{codesie}{Code SANDRE de la station concernée}

\item{codeoperation}{Code SANDRE de l'opération concernée}

\item{communeinsee}{Code INSEE de la commune (\code{vide} par défaut)}

\item{bbox}{Zone géographique concernée, au format bbox généré  par sf, en WGS84/EPSG 4326}

\item{date_start}{Date de début du prélèvement d'échantillon : \code{NA} aucune (par défaut) - Format \code{AAAA-MM-JJ}}

\item{date_end}{Date de fin du prélèvement d'échantillon : \code{NA} aucune (par défaut) - Format \code{AAAA-MM-JJ}}
}
\description{
Cette fonction permet de collecter les données d'opérations de suivi physico-chimique depuis Hub'eau
}
\examples{
PC.hubeau.operation("06400085")
PC.hubeau.operation(codesie_sandre)
PC.hubeau.operation("V2205003", date_start = "2019-01-01", date_end = "2020-01-01")
}
