% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.lac.vuespatiale.R
\name{PC.lac.vuespatiale}
\alias{PC.lac.vuespatiale}
\title{Création de vues spatialisée de profils verticaux lacustres}
\usage{
PC.lac.vuespatiale(
  data,
  contourplandeau,
  profondeur = c("zmax", "2"),
  legendemin = NA_real_,
  legendemax = NA_real_,
  facetting = NA_character_,
  save = F,
  projet = NA_character_,
  format = ".png"
)
}
\arguments{
\item{data}{Jeu de données à représenter, issu de fd_production.physicochimie_mesures}

\item{contourplandeau}{Contour de lac au format sf}

\item{profondeur}{Profondeur à laquelle les données sont à représenter (\code{zmax} par défaut, sinon valeur libre positive à saisir)}

\item{legendemin}{Valeur minimale de la légende (valeur par défaut fonction du paramètre représenté)}

\item{legendemax}{Valeur maximale de la légende (valeur par défaut fonction du paramètre représenté)}

\item{facetting}{Variable sur laquelle réaliser un facetting (aucun par défaut)}

\item{save}{Si \code{FALSE} (par défault), n'enregistre pas les figures. Si \code{TRUE}, les enregistre.}

\item{projet}{Nom du projet}

\item{format}{Défini le format d'enregistrement (par défaut .png)}
}
\description{
Cette fonction permet de créer des vues spatialisée de profils verticaux lacustres de paramètres physico-chimiques
}
\examples{
PC \%>\% PC.lac.vuespatiale()
PC \%>\% PC.lac.vuespatiale(contourplandeau = ContourPlandeau)
PC \%>\% PC.lac.vuespatiale(contourplandeau = ContourPlandeau, facetting = "pcmes_date")
PC \%>\% PC.lac.vuespatiale(contourplandeau = ContourPlandeau, facetting = "pcmes_date", profondeur = "zmax", legendemin = 0, legendemax = 100)
}
\keyword{physico-chimie}
