% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.parametres.R
\name{PC.parametres}
\alias{PC.parametres}
\title{Reformatage les paramètres physico-chimiques}
\arguments{
\item{data}{Jeu de données contenant une colonne Parametre ou ParametreNom}
}
\description{
Reformatage les paramètres physico-chimiques avec les noms uniformisés, les codes SANDRE et les unités
}
\examples{
PC.parametres(data)
}
\keyword{donnees}
