% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.renommage.R
\name{PC.renommage}
\alias{PC.renommage}
\title{Renommage des champs de données physico-chimiques issues de Hub'eau}
\usage{
PC.renommage(data)
}
\arguments{
\item{data}{Jeu de données issu de Hub'eau}
}
\description{
Renommage des champs de données physico-chimiques issues de Hub'eau vers le format de la base de données interne
}
\examples{
PC.renommage(data)
}
\keyword{donnees}
