% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.variables.renommage.R
\name{PC.variables.renommage}
\alias{PC.variables.renommage}
\title{Renommage des variables de physico-chimie}
\usage{
PC.variables.renommage(
  data = data,
  formatentree = c("Tous", "pcmes", "pcsvi"),
  formatsortie = c("chmes", "pcsvi", "param")
)
}
\arguments{
\item{data}{Data.frame contenant des variables à renommer}

\item{formatentree}{Type de format de noms de champs en entrée : \code{Tous} (par défaut), ou bien uniquement pour convertir certains : \code{pcmes} et \code{pcsvi}}

\item{formatsortie}{Type de format de noms de champs en sortie : \code{pcmes} (par défaut), \code{pcsvi} ou \code{param} (pour utilisation générique)}
}
\description{
Cette fonction permet de renommer des variables de traitement de données de physico-chimie
}
\examples{
Resultats \%>\% PC.variables.renommage()
Resultats \%>\% PC.variables.renommage(formatsortie = "param")
}
\keyword{chronique}
