% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personnel.projet.R
\name{personnel.projet}
\alias{personnel.projet}
\title{Extraction des données de temps de travail par projet}
\usage{
personnel.projet(projet = "Étude Leue")
}
\description{
Extrait au format excel les données plus ou moins détaillées des coûts du personnel par projet
}
\examples{
personnel.projet("Étude Leue")
}
\keyword{personnel}

