% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.brut.R
\name{poissons.brut}
\alias{poissons.brut}
\title{Exportation des résultats bruts de pêche}
\usage{
poissons.brut(station = "SAI2-3", date = "2014-09-10")
}
\arguments{
\item{station}{Code de la station}

\item{date}{Date de la pêche}
}
\description{
Cette fonction permet d'exporter les résultats bruts et élaborés de pêche au format excel
}
\examples{
poissons.brut("SOR10-2", "2015-05-19")
}
\keyword{poissons}

