% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.CAA.R
\name{poissons.CAA}
\alias{poissons.CAA}
\title{Collecte des données de CAA et CAR pour une station donnée}
\description{
Cette fonction permet de synthétiser les données de cotes d'abondance attendues et observées pour une station donnée
}
\examples{
poissons.CAA(data)
}
\keyword{poissons}
