% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.CAAvue.R
\name{poissons.CAAvue}
\alias{poissons.CAAvue}
\title{Affichage de graphes piscicoles}
\description{
Cette fonction permet de créer des comparaisons entre les espèces attendues et observées. Le jeu de données doit en entrée doit être issu de poissons.CAA
}
\examples{
poissons.CAAvue(data)
poissons.CAAvue(data, export=T)
poissons.CAA("MAD6-2") \%>\% poissons.CAAvue()
poissons.CAA("MAD6-2") \%>\% poissons.CAAvue(export = T)
}
\keyword{poissons}
