% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.captures.BDD.R
\name{poissons.captures.BDD}
\alias{poissons.captures.BDD}
\title{Extraction de l'ensemble des données de captures de poissons}
\usage{
poissons.captures.BDD()
}
\description{
Récupère l'ensemble des données de captures de poissons de Maxifish dans un dataframe
}
\examples{
poissons.captures.BDD()
}
\keyword{donnees}

