% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.classes.R
\name{poissons.classes}
\alias{poissons.classes}
\title{Calcul de classes de taille piscicoles}
\usage{
poissons.classes(Captures = Captures)
}
\arguments{
\item{Captures}{Dataframe contenant les captures}
}
\description{
Cette fonction permet de calculer les classes de taille des espèces piscicoles
}
\examples{
poissons.classes(Captures)
Captures <- poissons.classes(Captures)
}
\keyword{poissons}
