% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.contexte.R
\name{poissons.contexte}
\alias{poissons.contexte}
\title{contexte de données piscicoles}
\usage{
poissons.contexte(data = data)
}
\arguments{
\item{data}{Data.frame issu de la fonction \code{poissons.operations}}
}
\description{
Cette fonction permet d'établir le contexte de données piscicoles
}
\examples{
poissons.operations(data.frame(Nom = "SOR10-2"), CodeOperation = TRUE, Sortie = "Complet") \%>\% poissons.contexte()
}
\keyword{poissons}
