% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.cpue.R
\name{poissons.cpue}
\alias{poissons.cpue}
\title{Calcul des CPUE de captures}
\usage{
poissons.cpue(captures, placettes)
}
\arguments{
\item{captures}{Données de captures (format Multifish issu de \code{poissons.captures})}

\item{placettes}{Données de placettes (format Multifish issu de \code{poissons.placettes})}
}
\value{
\code{contexte_operations}, \code{effort_peche_filets_brut}, \code{effort_peche_filets_synthese}, \code{captures_n}, \code{captures_b}, \code{cpue_n}, \code{cpue_b}
}
\description{
Cette fonction permet de calculer les CPUE de captures en grands milieux
}
\examples{
poissons.cpue(captures, placettes)
}
\keyword{poissons}
