% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.expertise.R
\name{poissons.expertise}
\alias{poissons.expertise}
\title{Exportation des avis à dire d'expert}
\usage{
poissons.expertise(station = "SOR10-2", date = "2015-05-19")
}
\arguments{
\item{station}{Code de la station}

\item{date}{Date de la pêche}
}
\description{
Cette fonction permet d'exporter les avis à dire d'expert des opérations de suivi piscicole
}
\examples{
poissons.expertise("SOR10-2", "2015-05-19")
}
\keyword{poissons}
