% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.fiche.R
\name{poissons.fiche}
\alias{poissons.fiche}
\title{Création de fiches de résultats de pêche}
\arguments{
\item{station}{Code de la station}

\item{date}{Date de la pêche}

\item{commentaires}{\code{FALSE} par défault}
}
\description{
Cette fonction permet de créer une fiche de synthèse des résultats de pêche
}
\examples{
poissons.fiche("AIN18-4", "2013-09-27", commentaires = FALSE)
}
\keyword{poissons}
