% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.histogramme.R
\name{poissons.histogramme}
\alias{poissons.histogramme}
\title{Affichage d'histogrammes taille-effectif}
\arguments{
\item{Captures}{Dataframe contenant les captures}
}
\description{
Cette fonction permet de réaliser des histogrammes taille-effectif de poissons
}
\examples{
poissons.histogramme(Captures)
}
\keyword{poissons}
