% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.inventaires.R
\name{poissons.inventaires}
\alias{poissons.inventaires}
\title{Listing des inventaires pour une station}
\usage{
poissons.inventaires(station = "")
}
\arguments{
\item{station}{Code de la station}
}
\description{
Cette fonction permet de lister les différents inventaires pour une station donnée ou l'ensemble de la BDD
}
\examples{
poissons.inventaires()
poissons.inventaires("SOR10-2")
}
\keyword{poissons}
