% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.lac.irsteaversteleos.R
\name{poissons.lac.irsteaversteleos}
\alias{poissons.lac.irsteaversteleos}
\title{Transformation du format OFB/INRAE vers format Teleos}
\usage{
poissons.lac.irsteaversteleos(data)
}
\arguments{
\item{data}{Fichier en entrée (format OFB/INRAE)}
}
\description{
Cette fonction permet de transformer les données de suivis piscicoles de lacs du format OBF/INRAE vers celui Teleos pour import dans Multifish
}
\examples{
poissons.lac.irsteaversteleos("NAS-JB/Études/2016_Ilay-Val/Résultats/Poissons/Résultats_ONEMA/iLAY_2016_irstea_brut.xlsx")
}
\keyword{poissons}
