% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.NTT.R
\name{poissons.NTT}
\alias{poissons.NTT}
\title{Calcul des NTT}
\arguments{
\item{Données}{Dataframe issu de poissons.stations()}
}
\description{
Cette fonction permet de calculer les Niveaux Typologiques Théoriques
}
\examples{
poissons.NTT(station)
}
\keyword{poissons}
