% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.placettes.R
\name{poissons.placettes}
\alias{poissons.placettes}
\title{Listing des opérations pour une station}
\usage{
poissons.placettes(station = "", complet = FALSE)
}
\arguments{
\item{station}{Code de la station}

\item{complet}{Données simplifiées \code{FALSE} (par défault) ou complètes \code{TRUE}}
}
\description{
Cette fonction permet de lister les différentes opérations pour une station donnée
}
\examples{
poissons.placettes()
poissons.placettes("SOR10-2")
poissons.placettes(complet=TRUE)
poissons.placettes("SOR10-2",complet=TRUE)
}
\keyword{poissons}
