% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.plansdeau.ofbversteleos.R
\name{poissons.plansdeau.ofbversteleos}
\alias{poissons.plansdeau.ofbversteleos}
\title{Transformation du format OFB/INRAE vers format Teleos}
\usage{
poissons.plansdeau.ofbversteleos(data)
}
\arguments{
\item{data}{Fichier en entrée (format OFB/INRAE)}
}
\description{
Cette fonction permet de transformer les données de suivis piscicoles de lacs du format OFB/INRAE vers celui Teleos pour import dans Multifish
}
\examples{
poissons.plansdeau.ofbversteleos("NAS-JB/Études/2016_Ilay-Val/Résultats/Poissons/Résultats_ONEMA/iLAY_2016_irstea_brut.xlsx")
}
\keyword{poissons}
