% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.poids.R
\name{poissons.poids}
\alias{poissons.poids}
\title{Calcul des poids attendus en fonction de la taille pour un jeu de captures}
\usage{
poissons.poids(data = data, data_a_completer = NA_character_, pas_taille = 25)
}
\arguments{
\item{data}{Jeu de données de référence, issu de \code{poissons.captures}}

\item{data_a_completer}{Jeu de données à compléter (optionnel, si absent, sortie d'un tableau de référence), issu ou bien au format de \code{poissons.captures}}

\item{pas_taille}{Pas de taille (25 mm par défaut) des estimations produites. Ce paramètre est ignoré dans le cas du complément d'un jeu de données fourni via le paramètre \code{data_a_completer}.}
}
\description{
Calcule, à partir d'une jeu de données de captures, les poids attendus pour des tailles données
}
\examples{
poissons.poids(Captures)
poissons.poids(Captures, pas_taille = 10)
jeu_donnees_reference \%>\%
   filter(codeespece \%in\% liste_especes_traitees$codeespece) \%>\%
   group_split(codeespece) \%>\% 
   map_dfr(~ poissons.poids(data = ., data_a_completer = data_brutes))
}
\keyword{donnees}
