% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.resultats.BDD.R
\name{poissons.resultats.BDD}
\alias{poissons.resultats.BDD}
\title{Extraction des données des résultats de Maxifish}
\arguments{
\item{periode}{Permet de limiter la durée des données traitées : 10ans (par défaut), 20ans, Complet , 4campagnes}
}
\description{
Récupère l'ensemble des données de résultats de Maxifish dans un dataframe
}
\examples{
poissons.resultats.BDD()
}
\keyword{donnees}
