% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.resultats.R
\name{poissons.resultats}
\alias{poissons.resultats}
\title{Extraction des résultats pour une station}
\arguments{
\item{ListeStations}{Dataframe contenant un colonne "Nom" avec le code de la station (RHJ)}

\item{Sortie}{Forme du dataframe de sortie - \code{Simple} (par défault), \code{Propre} (format diffusable, avec stations) ou \code{Complet} (tous les champs)}
}
\description{
Cette fonction permet de collecter les différents résultats pour une liste donnée de stations
}
\examples{
poissons.resultats()
Resultats <- poissons.resultats()
poissons.resultats(data.frame(Nom = "SOR10-2"), Sortie = "Propre")
}
\keyword{donnees}
