% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.traitement.R
\name{poissons.traitement}
\alias{poissons.traitement}
\title{Traitement semi-automatisé de données piscicoles}
\description{
Cette fonction permet de traiter de manière semi-automatisée des données piscicoles (histogrammes, graphiques CAA/CAR, etc.)
}
\examples{
poissons.traitement(data)
poissons.traitement(data, export=T)
poissons.captures("AIN18-4") \%>\% poissons.traitement(export=T)
}
\keyword{poissons}
