% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profil.lac.R
\name{profil.lac}
\alias{profil.lac}
\title{Création de profils verticaux lacustres}
\usage{
profil.lac(
  PC,
  param = c("O2mg", "O2pourc", "cond", "ph", "temp", "redox", "chlorophylles",
    "phycocyanines", "secchi")
)
}
\arguments{
\item{param}{Paramètre physico-chimique à représenter (O2mg, O2pourc, cond, ph, temp, redox, chlorophylles, phycocyanines, secchi)}
}
\description{
Cette fonction permet de créer des profils graphiques de paramètres physico-chimiques
}
\examples{
profil.lac(PC, param = "O2mg")
PC \%>\% profil.lac(param = "redox")
}
