% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projet.calculInitial.R
\name{projet.calculInitial}
\alias{projet.calculInitial}
\title{Calcul initial des coûts de projet}
\arguments{
\item{NomProjet}{Nom du dataframe contenant les données "attendues" à traiter, dans le format de RecapTpsW}
}
\description{
Cette fonction permet de calculer les coûts d'un projet au moment de son montage. Elle s'appuie sur le format saisi dans la table RecapTpsW et réalise les conversions à partir des volumes unitaires, et ordonne les sujets
}
\examples{
RecapDataToAdd <- projet.calculInitial(NomProjet)
}
