% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapl.ouverture.adhesions.R
\name{sapl.ouverture.adhesions}
\alias{sapl.ouverture.adhesions}
\title{Ouverture des fichiers d'adhérents extraits de cartedepeche.fr}
\usage{
sapl.ouverture.adhesions(fichier)
}
\arguments{
\item{fichier}{Fichier à ouvrir}
}
\description{
Cette fonction permet l'ouverture des fichiers d'adhérents issus du site internet cartedepeche.fr
}
\examples{
sapl.ouverture.adhesions(fichier)
}
\keyword{stations}
