% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalimetrie.formatage.retrocalcul.R
\name{scalimetrie.formatage.retrocalcul}
\alias{scalimetrie.formatage.retrocalcul}
\title{Mise en forme des mesures de scalimétrie pour rétro-calcul avec RFishBC}
\usage{
scalimetrie.formatage.retrocalcul(
  lectures = NA_character_,
  mesures = NA_character_
)
}
\arguments{
\item{lectures}{Lectures des âges à traiter}

\item{mesures}{Mesures des stries de croissance à traiter}
}
\description{
Cette fonction permet de mettre en forme les données de scalimétrie en prévision d'un rétro-calcul d'âge avec RFishBC
}
\examples{
lectures \%>\% scalimetrie.formatage.retrocalcul(mesures)
}
\keyword{poissons}
\keyword{scalimétrie}
