% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalimetrie.import.R
\name{scalimetrie.import}
\alias{scalimetrie.import}
\title{Import des mesures de scalimétrie issues du logiciel Mosaic}
\usage{
scalimetrie.import(file = NA_character_)
}
\arguments{
\item{file}{Fichier à lire}
}
\description{
Cette fonction permet d'importer et de formater les mesures de scalimétrie réalisée à l'aide du logiciel Mosaic de la loupe binoculaire avec caméra
}
\examples{
mesures <- fs::dir_ls(data_dir, regexp = glue("\\\\{extension}$")) \%>\% map(~ scalimetrie.import(.)) \%>\% list_rbind()
}
\keyword{poissons}
\keyword{scalimétrie}
