% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sddlp.territoires.R
\name{sddlp.territoires}
\alias{sddlp.territoires}
\title{Jointure des territoires SDDLP aux SAPL}
\arguments{
\item{data}{Jeu de données contenant une colonne plosapl_sapl}
}
\description{
Cette fonction permet d'associer le territoire au sens du SDDLP à chaque SAPL
}
\examples{
sddlp.territoires(data)
sapl <- sf::st_read(dbD, query = "select * from fd_production.pecheloisir_sapl;") \%>\%  sddlp.territoires()
}
\keyword{sddlp}
