% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG.conversion.31n2l93.R
\name{SIG.conversion.31n2l93}
\alias{SIG.conversion.31n2l93}
\title{Conversion de coordonnées en 31N vers Lambert 93}
\usage{
SIG.conversion.31n2l93(coord_x = NA, coord_y = NA)
}
\arguments{
\item{coord_x}{Coordonnée x à convertir}

\item{coord_y}{Coordonnée y à convertir}
}
\description{
Cette fonction permet de convertir des coordonnées en 31N (32631) vers Lambert 93 (2154)
}
\examples{
SIG.conversion.31n2l93(695737.448, 5174080.793) \%>\% mapview::mapview()
}
\keyword{SIG}
