% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG.conversion.32n2l93.R
\name{SIG.conversion.32n2l93}
\alias{SIG.conversion.32n2l93}
\title{Conversion de coordonnées en 32N vers Lambert 93}
\usage{
SIG.conversion.32n2l93(coord_x = NA, coord_y = NA)
}
\arguments{
\item{coord_x}{Coordonnée x à convertir}

\item{coord_y}{Coordonnée y à convertir}
}
\description{
Cette fonction permet de convertir des coordonnées en 32N (32632) vers Lambert 93 (2154)
}
\examples{
SIG.conversion.32n2l93(276412.375, 5154007.570)
}
\keyword{SIG}
