% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG.conversion.dec2l93.R
\name{SIG.conversion.dec2l93}
\alias{SIG.conversion.dec2l93}
\title{Conversion de coordonnées en degrés décimaux vers Lambert 93}
\usage{
SIG.conversion.dec2l93(coord_x = NA, coord_y = NA)
}
\arguments{
\item{coord_x}{Coordonnée x à convertir}

\item{coord_y}{Coordonnée y à convertir}
}
\description{
Cette fonction permet de convertir des coordonnées en degrés décimaux (4326) vers du Lambert 93 (2154)
}
\examples{
SIG.conversion.dec2l93(5.972321, 46.806626) \%>\% mapview::mapview()
}
\keyword{SIG}
