% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG.conversion.dms2dec.R
\name{SIG.conversion.dms2dec}
\alias{SIG.conversion.dms2dec}
\title{Conversion de coordonnées en degrés-minutes-secondes vers degrés décimaux}
\usage{
SIG.conversion.dms2dec(
  dms,
  separators = c("º", "°", "’’", "''", "'", "’", "\\"", "\\\\?")
)
}
\arguments{
\item{dms}{Coordonnées à convertir}

\item{separators}{Séparateurs de valeurs utilisés}
}
\description{
Cette fonction permet de convertir des coordonnées en degrés-minutes-secondes vers des degrés décimaux
}
\examples{
SIG.conversion.dms2dec("41° 34.1826' N")
}
\keyword{SIG}
