% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG.conversion.l932dec.R
\name{SIG.conversion.l932dec}
\alias{SIG.conversion.l932dec}
\title{Conversion de coordonnées en Lambert 93 vers degrés décimaux}
\usage{
SIG.conversion.l932dec(coord_x = NA, coord_y = NA)
}
\arguments{
\item{coord_x}{Coordonnée x à convertir}

\item{coord_y}{Coordonnée y à convertir}
}
\description{
Cette fonction permet de convertir des coordonnées en du Lambert 93 (2154) vers degrés décimaux (4326)
}
\examples{
SIG.conversion.l932dec(888227, 6655627) \%>\% mapview::mapview()
}
\keyword{SIG}
