% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG.conversion.l93dec2.R
\name{SIG.conversion.l93dec2}
\alias{SIG.conversion.l93dec2}
\title{Conversion de coordonnées en Lambert 93 vers degrés décimaux}
\usage{
SIG.conversion.l93dec2(coord_x, coord_y)
}
\arguments{
\item{coord_x}{Coordonnée x à convertir}

\item{coord_y}{Coordonnée y à convertir}
}
\description{
Cette fonction permet de convertir des coordonnées en du Lambert 93 vers degrés décimaux
}
\examples{
SIG.conversion.l93dec2(888227, 6655627)
}
\keyword{SIG}
