% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG.exportSHP.R
\name{SIG.exportSHP}
\alias{SIG.exportSHP}
\title{Export de données en shp}
\usage{
SIG.exportSHP(Data = Stations, XLambert = XLambert, YLambert = YLambert,
  nomfichier = "Export_stations")
}
\arguments{
\item{Data}{Dataframe à exporter}

\item{XLambert}{Colonne des X}

\item{YLambert}{Colonne des Y}

\item{nomfichier}{Nom du fichier de sortie}
}
\description{
Cette fonction permet d'exporter un shapefile en Lambert 93
}
\examples{
SIG.exportSHP(Stations, XLambert, YLambert, "Export_stations")
}
\keyword{poissons}
\keyword{stations}

