% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG.flux.wfs.R
\name{SIG.flux.wfs}
\alias{SIG.flux.wfs}
\title{Collecte de données à partir d'un flux WFS}
\usage{
SIG.flux.wfs(
  url = NA_character_,
  dataset = NA_character_,
  champ = NA_character_,
  valeur = NA_character_
)
}
\arguments{
\item{url}{URL de base du flux - Exemple : \code{https://data.geopf.fr/wfs/}}

\item{dataset}{Jeu de données recherché - Exemple : \code{ADMINEXPRESS-COG.LATEST:departement}}

\item{champ}{Champ concerné si recherche d'une valeur particulière - Exemple : \code{insee_dep}}

\item{valeur}{Valeur concernée si recherche d'une valeur particulière - Exemple : \code{39}}

\item{X}{Coordonnée X à rechercher (Lambert 93)}

\item{Y}{Coordonnée Y à rechercher (Lambert 93)}
}
\description{
Cette fonction permet de collecter des données à partir d'un flux WFS
}
\examples{
SIG.flux.wfs("https://data.geopf.fr/wfs/", "ADMINEXPRESS-COG.LATEST:departement")
SIG.flux.wfs("https://data.geopf.fr/wfs/", "ADMINEXPRESS-COG.LATEST:departement", "insee_dep", "39")
}
