% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG.style.qgis.R
\name{SIG.style.qgis}
\alias{SIG.style.qgis}
\title{Conversion de palette QGIS vers palette R}
\usage{
SIG.style.qgis(data, format = c("palette", "dataframe"))
}
\arguments{
\item{data}{Palette à convertir, au format xml}

\item{format}{Format de sortie : \code{palette} (par défaut) ou \code{dataframe}}
}
\description{
Cette fonction permet de convertir des palettes QGIS au format .sld vers une palette de couleur au format R
}
\examples{
SIG.style.qgis(palette)
tbl(dbD, dbplyr::in_schema("public", "layer_styles")) \%>\% filter(f_table_name == "topographie_habitats") \%>\% collect() \%>\% SIG.style.qgis(.$stylesld)
}
\keyword{SIG}
