% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soleil.horaires.R
\name{soleil.horaires}
\alias{soleil.horaires}
\title{Calcul des horaires de soleil}
\usage{
soleil.horaires(
  latitude = NA_real_,
  longitude = NA_real_,
  date_start = NA_character_,
  date_end = NA_character_
)
}
\arguments{
\item{latitude}{Latitude concernée}

\item{longitude}{Longitude concernée}

\item{date_start}{Date de début de la période concernée}

\item{date_end}{Date de fin de la période concernée}
}
\description{
Fait le calcul des horaires de présence de soleil, dont horaires autorisés de pêche
}
\examples{
soleil.horaires(48.85331560269666, 2.3456499560163597, "2026-01-01", "2026-12-31")
soleil.horaires(46.84015684582083, 5.7054853214538115, "2026-01-01", "2026-12-31")
}
