% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.CodeRDT.R
\name{stations.CodeRDT}
\alias{stations.CodeRDT}
\title{Extraire la distance à la source ou le code écosystème du CodeRDT}
\usage{
stations.CodeRDT(data = data, DistSource = T)
}
\arguments{
\item{data}{Dataframe contenant une colonne CodeRDT}

\item{DistSource}{Si \code{T} (par défaut), extrait la distance à la source - Extrait sinon le code de l'écosystème}
}
\description{
Cette fonction permet d'extraire la distance à la source ou le code écosystème à partir du CodeRDT
}
\examples{
stations.CodeRDT(data)
stations.CodeRDT(data, DistSource = F)
arrange(CodeRDT, Date) \%>\% stations.CodeRDT(DistSource = F)
}
\keyword{stations}
