% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.coderhj.R
\name{stations.coderhj}
\alias{stations.coderhj}
\title{Extraire la distance à la source ou le code écosystème du coderhj}
\arguments{
\item{data}{Dataframe contenant une colonne coderhj}

\item{DistSource}{Si \code{T} (par défaut), extrait la distance à la source - Extrait sinon le code de l'écosystème}
}
\description{
Cette fonction permet d'extraire la distance à la source ou le code écosystème à partir du coderhj
}
\examples{
stations.coderhj(data)
stations.coderhj(data, DistSource = F)
arrange(coderhj, Date) \%>\% stations.coderhj(DistSource = F)
}
\keyword{stations}
