% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.DistSource.R
\name{stations.DistSource}
\alias{stations.DistSource}
\title{Extraire la distance à la source du CodeRDT}
\usage{
stations.DistSource(data = data)
}
\arguments{
\item{data}{Dataframe contenant une colonne CodeRDT}
}
\description{
Cette fonction permet d'extraire la distance à la source à partir du CodeRDT
}
\examples{
stations.DistSource(data)
}
\keyword{stations}
