% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.ecosysteme.R
\name{stations.ecosysteme}
\alias{stations.ecosysteme}
\title{Listage des stations d'un cours d'eau}
\arguments{
\item{Nom}{du cours d'eau}

\item{shp}{Si \code{FALSE} (par défault), n'exporte pas de shp dans le répertoire courant.}
}
\description{
Cette fonction permet de lister les stations d'un cours d'eau donné dans les BDD poisson et chronique
}
\examples{
stations.ecosysteme()
stations.ecosysteme("Ain")
stations.ecosysteme("Ain", shp=T)
}
\keyword{stations}
