% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations}
\alias{stations}
\title{Extraction de données de stations}
\usage{
stations(station = "SOR10-2")
}
\arguments{
\item{Nom}{de la station}
}
\description{
Cette fonction permet d'extraire les données complètes d'une station
}
\examples{
stations("SOR10-2")
}
\keyword{stations}

