% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.territoire.R
\name{stations.territoire}
\alias{stations.territoire}
\title{Permet d'établir un périmètre afin d'en retenir ensuite les stations}
\arguments{
\item{Echelle}{Échelle spatiale de la synthèse (commune, canton, communauté de communes, département, région, contexte de PDPG, Hydro-écorégion, entité GEMAPI, Maître d'ouvrage, Milieu, Bassin versant, Sous-bassin versant, Polygone autre)}

\item{Territoire}{Territoire concerné. Unique ou sous forme de vecteur (c("Villerserine", "Villers-Robert") par exemple)}

\item{Liste}{\code{FALSE} par défault. Permet d'afficher une liste dans laquelle choisir le territoire concerné}
}
\description{
Cette fonction permet d'extraire sous forme spatiale (sf) un territoire à une échelle donnée afin de le réutiliser par croisement spatial dans les fonctions chronique.stations ou poissons.synthese par exemple
}
\examples{
stations.territoire(data)
stations.territoire("ComCom", "200071595")
stations.territoire("Commune", c("Villerserine", "Villers-Robert"))
stations.territoire("Commune", c("39568", "39571"))
stations.territoire("Commune", c("Villerserine", "39571"))
stations.territoire("Milieu", c("Lac de Chalain", "Bonlieu"))
}
\keyword{stations}
