% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermie.analyse.R
\name{thermie.analyse}
\alias{thermie.analyse}
\title{Calcul de paramètres clés sur des chroniques de température}
\arguments{
\item{CodeRDT}{Nom de la station}

\item{Annee}{Année du suivi}

\item{jour}{Colonne du data.frame contenant les dates}

\item{temp}{Colonne du data.frame contenant les températures}

\item{pasdetemps}{Indique le pas de temps entre chaque mesure : 1 pour une heure}
}
\description{
Cette fonction permet de calculer des paramètres clés sur des chroniques de température
}
\examples{
thermie.analyse("DRO14-8","2013",jour=tab[,1],temp=tab[,3],tempmin=4,tempmax=19,tempmaxextreme=25,pasdetemps=1)
}
\keyword{température}
