% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermie.suivi.R
\name{thermie.suivi}
\alias{thermie.suivi}
\title{Extraction suivi terrain thermie}
\usage{
thermie.suivi(x = "ORA2-7", Type = c("MO", "Opérateurs", "Station", "Date",
  "Sonde"))
}
\arguments{
\item{Type}{Type de donnée de suivi}
}
\description{
Cette fonction permet d'extraire des données de suivi de terrain
}
\examples{
thermie.suivi("CG39", Type = "MO")
thermie.suivi("JB-Stéphane", Type = "Opérateurs")
thermie.suivi("DRO14-2", Type = "Station")
thermie.suivi("2015-10-23", Type = "Date")
thermie.suivi("9759803", Type = "Sonde")
}

