% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topographie.correction.z.R
\name{topographie.correction.z}
\alias{topographie.correction.z}
\title{Correction altimétrique de coordonnées}
\usage{
topographie.correction.z(leves = NA_character_, correction = 0)
}
\arguments{
\item{leves}{Nuage de points issu de \code{topographie.mesures}}

\item{correction}{Valeur de la correction à apporter, en mètres}
}
\description{
Cette fonction permet de corriger la coordonnée Z de points au format \code{sf}
}
\examples{

leves_1 \%>\% topographie.correction.z(0.1)
}
\keyword{topographie}
