% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topographie.figure.profil.automatique.R
\name{topographie.figure.profil.automatique}
\alias{topographie.figure.profil.automatique}
\title{Création automatisée de vues de profils le long de transects}
\usage{
topographie.figure.profil.automatique(
  transects = NA,
  leves = NA,
  buffer = 10,
  save = F,
  save_name = NA_character_
)
}
\arguments{
\item{transects}{Dataframe avec les identifiants et les coordonnées de début et de fin des transects à représenter}

\item{leves}{Nuage de points issu de \code{topographie.mesures}}

\item{buffer}{Distance maximale entre les points et le transect (\code{10} par défaut)}

\item{save}{Si \code{FALSE} (par défaut), n'enregistre pas les figures. Si \code{TRUE}, les enregistre.}

\item{save_name}{Nom de fichier à utiliser pour l'enregistrement. \code{"{today()}_Vue_profil.png"} si vide.}
}
\description{
Cette fonction permet de représenter des profils (topographiques par exemple) le long de transects
}
\examples{
transect \%>\% topographie.figure.profil.automatique()
}
\keyword{topographie}
