% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topographie.figure.profil.R
\name{topographie.figure.profil}
\alias{topographie.figure.profil}
\title{Vue d'un profil le long d'un transect}
\usage{
topographie.figure.profil(
  data = NA_character_,
  etiquette = T,
  pointsprojetes = T,
  save = F
)
}
\arguments{
\item{data}{Jeu de données en sortie de topographie.profil.raster()}

\item{etiquette}{Affichage des étiquettes des points d'intérêt : \code{TRUE} (par défaut) ou \code{FALSE}}

\item{pointsprojetes}{Affichage des points d'intérêt projetés le long du transect : \code{TRUE} (par défaut) ou \code{FALSE}}

\item{save}{Si \code{FALSE} (par défaut), n'enregistre pas les figures. Si \code{TRUE}, les enregistre.}
}
\description{
Cette fonction permet de représenter un profil (topographique par exemple) d'un transect
}
\examples{
data \%>\% topographie.figure.profil()
topographie.figure.profil(data, etiquette = F, pointsprojetes = F)
data \%>\% topographie.figure.profil(save = T)
topographie.figure.profil(raster = mon_raster_regroupe, transect = StationsTransect, points_projetes = Stations, points_projetes_position = "intermediaire") \%>\% topographie.figure.profil()
}
\keyword{topographie}
