% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topographie.figure.raster.transect.R
\name{topographie.figure.raster.transect}
\alias{topographie.figure.raster.transect}
\title{Vue d'un transect sur un raster}
\usage{
topographie.figure.raster.transect(
  raster = NA_character_,
  transect = NA_character_,
  points_projetes = NA_character_,
  lignes_projetees_sur_transect = NA_character_,
  etiquette = T,
  angle_etiquette = 0,
  position_fleche_nord = c("bl", "tl", "tr", "br"),
  facteur_largeur = 0.4,
  facteur_hauteur = 0.4,
  valeur_min = NA_real_,
  valeur_max = NA_real_,
  save = F
)
}
\arguments{
\item{raster}{Raster sur lequel on souhaite représenter le transect}

\item{transect}{Transect à représenter, en sortie de topographie.transect()}

\item{points_projetes}{Points d'intérêt à projeter le long du transect}

\item{lignes_projetees_sur_transect}{Transects de projection des points à représenter, en sortie de topographie.transect()}

\item{etiquette}{Affichage des étiquettes des points d'intérêt : \code{TRUE} (par défaut) ou \code{FALSE}}

\item{angle_etiquette}{Angle de rotation à appliquer aux étiquettes (0 par défaut)}

\item{position_fleche_nord}{Position de la flèche du nord (\code{"bl"} par défaut)}

\item{facteur_largeur}{Facteur d'extension du raster en largeur par rapport aux données (0.4 par défaut)}

\item{facteur_hauteur}{Facteur d'extension du raster en hauteur par rapport aux données (0.4 par défaut)}

\item{valeur_min}{Valeur minimale à représenter sur le raster et dans la légende (valeur minimale du raster par défaut)}

\item{valeur_max}{Valeur maximale à représenter sur le raster et dans la légende (valeur maximale du raster par défaut)}

\item{save}{Si \code{FALSE} (par défaut), n'enregistre pas les figures. Si \code{TRUE}, les enregistre.}
}
\description{
Cette fonction permet de représenter un transect sur un raster
}
\examples{
topographie.figure.raster.transect(raster, transect)
topographie.figure.raster.transect(raster, transect, points_projetes)
topographie.figure.raster.transect(raster, transect, points_projetes, lignes_projetees_sur_transect)
topographie.figure.raster.transect(raster, transect, points_projetes, lignes_projetees_sur_transect, facteur_largeur = 0.75, facteur_hauteur = 0.3)
topographie.figure.raster.transect(raster, transect, points_projetes, lignes_projetees_sur_transect, save = T)
}
\keyword{topographie}
