% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topographie.habitats.vue.R
\name{topographie.habitats.vue}
\alias{topographie.habitats.vue}
\title{Classes de valeurs de chroniques sous forme calendaire}
\usage{
topographie.habitats.vue(
  data = data,
  type = c("Substrats", "Vitesses", "Hauteurs", "Pôles"),
  proportions = TRUE,
  titre = "",
  caption = "FDPPMA39",
  position_echelle = "tr",
  ecoulement_orientation = 0,
  export = FALSE,
  format = ".png"
)
}
\arguments{
\item{data}{Objet au format sf contenant les données de substrats, hauteurs d'eau et vitesses de courant}

\item{type}{Type de représentation graphique : \code{Substrats}, \code{Vitesses} de courant, \code{Hauteurs} d'eau ou \code{Pôles} d'attraction}

\item{proportions}{Affichage des proportions absolues de chaque habitat : \code{TRUE} par défaut}

\item{titre}{Titre du graphique (vide par défaut)}

\item{caption}{Légende de bas de graphique : \code{Données FDPPMA39} par défaut}

\item{position_echelle}{Position de l'échelle (\code{tr} par défaut)}

\item{ecoulement_orientation}{Orientation par rapport au nord, en degrés, dans le sens des aiguilles d'une montre}

\item{export}{Si \code{TRUE}, exporte les différents résultats}

\item{format}{Défini le format d'enregistrement (par défaut .png)}
}
\description{
Cette fonction permet de représenter les classes de valeurs de référence pour des chroniques de mesures (température, niveaux, etc.) sous forme calendaire.
}
\examples{
topographie.habitats(36) \%>\% topographie.habitats.vue("Substrats")
topographie.habitats(tpiam_op_id) \%>\% topographie.habitats.vue("Vitesses")
topographie.habitats(tpiam_op_id) \%>\% topographie.habitats.vue("Hauteurs")
data(habitats_exemple) ; habitats_exemple \%>\% topographie.habitats.vue("Substrats", proportions = FALSE)
data(habitats_exemple) ; habitats_exemple \%>\% topographie.habitats.vue("Pôles", position_echelle = "br", caption = "Données CD39/FDPPMA39") + ggspatial::annotation_north_arrow(location = "tl")
}
\keyword{chronique}
