% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topographie.hypsometrie.R
\name{topographie.hypsometrie}
\alias{topographie.hypsometrie}
\title{Création d'une hypsométrie}
\usage{
topographie.hypsometrie(data = NA_character_)
}
\arguments{
\item{data}{Dataframe contenant une colonne \code{chsta_coord_z} ou une géométrie de type POINTZ. Affichage de différentes courbes si présence d'une colonne \code{plan_deau}}

\item{epsg}{EPSG à attribuer au}
}
\description{
Cette fonction permet de créer une vue hypsométrique
}
\examples{
data \%>\% topographie.hypsometrie()
}
\keyword{topographie}
