% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topographie.iam.R
\name{topographie.iam}
\alias{topographie.iam}
\title{Calcul d'un IAM}
\usage{
topographie.iam(
  data = NA_character_,
  operation = NA_character_,
  export = FALSE
)
}
\arguments{
\item{data}{Objet au format sf contenant les données de substrats, hauteurs d'eau et vitesses de courant}

\item{operation}{Dataframe contenant les caractéristiques de l'opération (date, cours d'eau, coordonnées, largeur moyenne, affluents/résurgences/sources, etc.)}

\item{export}{Si \code{TRUE}, exporte les différents résultats}
}
\description{
Cette fonction permet de calculer les résultats d'un IAM à partir des cartographies d'habitats
}
\examples{
iam <- topographie.iam(topographie.habitats("12"), iam.resultats("12"))
}
\keyword{topographie}
