% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topographie.profil.raster.R
\name{topographie.profil.raster}
\alias{topographie.profil.raster}
\title{Calcul d'un profil raster le long d'un transect}
\usage{
topographie.profil.raster(
  raster = NA_character_,
  epsg = 2154,
  transect = NA,
  points_projetes = tibble(var_name_1 = numeric()),
  points_projetes_position = c("lissee", "intermediaire", "reelle")
)
}
\arguments{
\item{raster}{Raster déjà importé et éventuellement regroupé ou localisation relative du raster}

\item{epsg}{EPSG du raster (\code{2154} par défaut)}

\item{transect}{Dataframe au format sf contenant 2 points ordonnées (départ -> arrivée) ou bien une LINESTRING (cours d'eau par exemple)}

\item{points_projetes}{Dataframe au format sf contenant les points que l'on souhaite projeter sur le profil}

\item{points_projetes_position}{Position souhaitée pour l'affichage des points remarquables : \code{lissee} (par défaut) : sur une valeur unique, \code{intermediaire} ou \code{reelle}}
}
\description{
Cette fonction permet de calculer le profil topographique ou autres le long d'un transect à partir d'un raster (MNT, etc.)
}
\examples{
topographie.profil.raster(raster = mon_raster_regroupe, transect = StationsTransect, points_projetes = Stations, points_projetes_position = "intermediaire") \%>\% magrittr::extract2(2)
topographie.profil.raster(raster = "Downloads/RGEALTI_FXX_0895_6595_MNT_LAMB93_IGN69.asc", transect = StationsTransect, points_projetes = Stations, points_projetes_position = "intermediaire") \%>\% magrittr::extract2(2)
}
\keyword{topographie}
