% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topographie.profil.R
\name{topographie.profil}
\alias{topographie.profil}
\title{Calcul d'un profil topographique le long d'un transect}
\usage{
topographie.profil(leves = NA_character_, transects = NA, buffer = 10)
}
\arguments{
\item{leves}{Nuage de points issu de `topographie.mesures`}

\item{transects}{Dataframe au format sf contenant 2 points ordonnées (départ -> arrivée) ou bien une LINESTRING (cours d'eau par exemple)}

\item{buffer}{Distance maximale entre les points et le transect}
}
\description{
Cette fonction permet de calculer le profil topographique ou autres le long d'un transect à partir d'un nuage de points
}
\examples{

topographie.profil(leves_1, transects, 150)
}
\keyword{topographie}
