% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topographie.transect.R
\name{topographie.transect}
\alias{topographie.transect}
\title{Création d'un transect}
\usage{
topographie.transect(data = NA_character_)
}
\arguments{
\item{data}{Dataframe contenant deux points - L'ordre de tri des points constituera le sens du transect}

\item{epsg}{EPSG à attribuer au transect}
}
\description{
Cette fonction permet de créer un transect à partir de deux points
}
\examples{
data \%>\% topographie.transect()
}
\keyword{topographie}
