% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troncon.recode.R
\name{troncon.recode}
\alias{troncon.recode}
\title{Recode note protocole tronçon}
\usage{
troncon.recode(troncon = data)
}
\arguments{
\item{data}{Chronique à valider}
}
\description{
Cette fonction permet de recoder les notes brutes du protocole tronçon en classes de qualité
}
\examples{
troncon.recode(data)
}
\keyword{data}
