% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDD.ecriture.R
\name{BDD.ecriture}
\alias{BDD.ecriture}
\title{Ouverture des bases de données pour écriture}
\arguments{
\item{Type}{Type de base de données}
}
\description{
Cette fonction permet de charger les bases de données de la fédération pour écrire dedans
}
\examples{
BDD.ecriture(Type = "Chroniques")
}
