% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDD.ouverture.R
\name{BDD.ouverture}
\alias{BDD.ouverture}
\title{Chargement des bases de données}
\usage{
BDD.ouverture(Type = c("Chroniques", "Poissons", "Macroinvertébrés",
  "Physico-chimie", "Temps de travail"))
}
\arguments{
\item{Type}{Type de base de données}
}
\description{
Cette fonction permet de charger les bases de données de la fédération
}
\examples{
BDD.ouverture(Type = "Chroniques")
}
