% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDD.ouverture.R, R/chronique.validation.R
\name{BDD.ouverture}
\alias{BDD.ouverture}
\title{Chargement des bases de données}
\usage{
BDD.ouverture(data)

BDD.ouverture(data)
}
\arguments{
\item{data}{Chronique à valider}

\item{Type}{Type de base de données}
}
\description{
Cette fonction permet de charger les bases de données de la fédération

Cette fonction permet d'obtenir des éléments afin de valider les chroniques
}
\examples{
BDD.ouverture(Type = "Thermie")
chronique.validation(data)
}

