% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDD.sauvegarde.R
\name{BDD.sauvegarde}
\alias{BDD.sauvegarde}
\title{Sauvegarde des bases de données}
\usage{
BDD.sauvegarde(x, Type = c("Thermie", "Physico-chimie", "Piézométrie",
  "Temps de travail"))
}
\arguments{
\item{Type}{Type de base de données}
}
\description{
Cette fonction permet de sauvegarder les bases de données de la fédération
}
\examples{
BDD.sauvegarde(Type = "Thermie")
}

