% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BV.ComByCoordL93.R
\name{BV.ComByCoordL93}
\alias{BV.ComByCoordL93}
\title{Recherche de caractéristiques administratives d'un point}
\usage{
BV.ComByCoordL93(X = as.numeric(0), Y = as.numeric(0))
}
\arguments{
\item{X}{Coordonnée X à rechercher (Lambert 93)}

\item{Y}{Coordonnée Y à rechercher (Lambert 93)}
}
\description{
Cette fonction permet de fournir les caractéristiques géographiques d'un point donné (commune, département, région, population, etc.) à partir de la GeoAPI
}
\examples{
BV.ComByCoordL93()
}
