% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.agregation.R
\name{chronique.agregation}
\alias{chronique.agregation}
\title{Agrégation de chroniques}
\usage{
chronique.agregation(data = data, export = FALSE)
}
\arguments{
\item{data}{Data.frame contenant a minima une colonne Date, une colonne Heure et une colonne Valeur}
}
\description{
Cette fonction permet d'agréger des chroniques de mesures (température, niveaux, etc.)
}
\examples{
DataTravail <- chronique.agregation(data)
DataTravail[[1]];DataTravail[[2]];DataTravail[[3]];DataTravail[[4]]
chronique.agregation(data, export= T) # Export sous forme de fichier excel avec 4 onglets
}
\keyword{chronique}
