% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.analyse.R
\name{chronique.analyse}
\alias{chronique.analyse}
\title{Analyse de chroniques}
\usage{
chronique.analyse(
  data = data,
  typemesure = c("Thermie", "Thermie barométrique", "Thermie piézométrique",
    "Barométrie", "Piézométrie", "Piézométrie brute", "Piézométrie compensée",
    "Piézométrie calée", "Piézométrie NGF", "Oxygénation", "Hydrologie",
    "Pluviométrie"),
  pasdetemps = 1,
  seuils = c(25, 22, 19, 15, 4),
  seuilexcesdefaut = 12,
  degresjours = F
)
}
\arguments{
\item{data}{Data.frame contenant a minima une colonne chmes_date, une colonne chmes_heure et une colonne chmes_valeur}

\item{typemesure}{Défini le type de données et modifie l'analyse en fonction}

\item{pasdetemps}{Indique le pas de temps entre chaque mesure : 1 pour une heure (1 par défaut)}

\item{seuils}{Seuils de valeurs (25,22,19,15,4 par défaut) dont il faut tester les dépassements. L'ordre de sortie correspond à l'ordre de cette liste.}

\item{seuilexcesdefaut}{Seuil en-deça duquel les dépassements sont testés par défaut, sinon ils sont testés par excès (valeur de 12 par défaut).}

\item{degresjours}{Calcul des degrés/jours (\code{FALSE} par défault)}
}
\description{
Cette fonction permet d'analyser des chroniques de mesures (température, niveaux, etc.)
}
\examples{
chronique.analyse(data, "Thermie")
}
\keyword{chronique}
