% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.analyse.R
\name{chronique.analyse}
\alias{chronique.analyse}
\title{Analyse de chroniques}
\usage{
chronique.analyse(data = data, pasdetemps = 1)
}
\arguments{
\item{data}{Data.frame contenant a minima une colonne Date, une colonne Heure et une colonne Valeur}

\item{pasdetemps}{Indique le pas de temps entre chaque mesure : 1 pour une heure (1 par défaut)}

\item{Titre}{Titre du graphique (vide par défaut)}

\item{legendeY}{Défini le texte de la légende de l'axe Y (Température (°C) par défaut)}

\item{duree}{Si \code{Complet} (par défault), affichage de l'année complète.  Si \code{Relatif}, affichage uniquement de la période concernée.}

\item{Vmm30j}{Si \code{FALSE} (par défault), n'affiche pas les
Vmm30j.  Si \code{TRUE}, les affiche.}

\item{Vminmax}{Si \code{TRUE} (par défault), affiche pas les
valeurs journalières minimales et maximales.  Si \code{FALSE}, ne les affiche pas.}

\item{Ymin}{Valeur minimale de l'axe des Y (0 par défaut)}

\item{Ymax}{Valeur maximale de l'axe des Y (aucune par défaut)}
}
\description{
Cette fonction permet d'analyser des chroniques de mesures (température, niveaux, etc.)
}
\examples{
chronique.analyse("ANG0-4", 2016, data, pasdetemps = 1)
chronique.analyse(as.character(Stations[i,2]), Annee, data, pasdetemps = 1)
}
\keyword{chronique}
