% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.cle.R
\name{chronique.cle}
\alias{chronique.cle}
\title{Créer d'une clé pour les chroniques}
\usage{
chronique.cle(
  data = data,
  anneebiologique = TRUE,
  formatcle = c("SAT", "SATD", "SA", "ST", "MA", "HST", "HT", "HS")
)
}
\arguments{
\item{data}{Data.frame issu de la fonction chronique.mesures ou chronique.resultats}

\item{anneebiologique}{Si \code{TRUE} (par défault), s'appuie sur l'année biologique lorsque le paramètre année est inclus à la clé}

\item{formatcle}{Format de la clé d'identification, avec coderhj (S), annee (A), typemesure (T), nbj (D), Milieu (M), Date-Heure (H) par exemple de la forme coderhj_annee_typemesure (SAT) (par défault), coderhj_annee_typemesure_nbj (SATD), coderhj_annee (SA), coderhj_typemesure (ST) ou milieu_annee (MA) ou time_coderhj_typemesure (HST) ou time_typemesure (HT) ou time_coderhj (HS)}
}
\description{
Cette fonction permet de créer une clé (champ Cle) pour les chroniques (mesures, résultats, etc.)
}
\examples{
Resultats \%>\% chronique.cle()
Resultats \%>\% chronique.cle(anneebiologique = F, formatcle = "SA")
}
\keyword{chronique}
