% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.compensation.barometrie.R
\name{chronique.compensation.barometrie}
\alias{chronique.compensation.barometrie}
\title{Compensation barométrique des séries piézométriques}
\usage{
chronique.compensation.barometrie(
  data,
  modalite_rattachement = c("Interpolation", "Proximité"),
  duree_max_rattachement = 1,
  sortie = c("compensé", "compensé_avec_vide", "tout", "large")
)
}
\arguments{
\item{data}{Chronique à modifier, avec un champ chmes_date, un champ chmes_heure et un champ chmes_valeur}

\item{modalite_rattachement}{Mode de rattachement barométrique : \code{Interpolation} (par défaut) ou \code{Proximité} temporelle}

\item{duree_max_rattachement}{Durée maximale de rattachement, en heures - \code{1} (par défaut)}

\item{sortie}{Format de sortie - \code{compensé} (par défaut) : uniquement le jeu de données nouvellement compensé, \code{compensé_avec_vide} : uniquement le jeu de données nouvellement compensé, \code{tout} : jeu de données initial + jeu de données compensées, \code{large} : format large pour étude intermédiaire.}
}
\description{
Permet de compenser barométriquement des séries piézométriques, avec la valeur barométrique la plus proche
}
\examples{
chronique.compensation.barometrie(data)
chronique.compensation.barometrie(data, "Proximité", 2)
}
\keyword{chronique}
