% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.complete.R
\name{chronique.complete}
\alias{chronique.complete}
\title{Complément de chroniques}
\arguments{
\item{data}{Data.frame contenant a minima une colonne Date, une colonne Heure et une colonne Valeur}
}
\description{
Cette fonction permet de compléter les chroniques présentant des valeurs manquantes
}
\examples{
chronique.complete(data)
}
\keyword{chronique}
