% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.decalage.R
\name{chronique.decalage}
\alias{chronique.decalage}
\title{Décalage de données brutes de chronique}
\usage{
chronique.decalage(data, decalage = 2598.581)
}
\arguments{
\item{data}{Chronique à modifier, avec un champ chmes_date, un champ chmes_heure et un champ chmes_valeur}

\item{decalage}{Valeur de décalage (en jours, positif pour avancer ou négatif pour reculer) à calculer par soustration des deux dates au format lubridate}
}
\description{
Permet de corriger des données brutes de chronique présentant des valeurs de date et heure aberrantes. ATTENTION : ne gère pas les ID
}
\examples{
ymd_hms("2017-07-05 11:40:00")-ymd_hms("2010-05-24 21:44:00")
chronique.decalage(data, decalage = 2598.581)
chronique.decalage(decalage = -57)
}
\keyword{chronique}
