% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.decalage.R
\name{chronique.decalage}
\alias{chronique.decalage}
\title{Décalage de données brutes de chronique}
\arguments{
\item{data}{Chronique à modifier, avec un champ Date, un champ Heure et un champ Valeur}

\item{decalage}{Valeur de décalage à calculer par soustration des deux dates au format lubridate}
}
\description{
Permet de corriger des données brutes de chronique présentant des valeurs de date et heure aberrantes. ATTENTION : ne gère pas les ID
}
\examples{
ymd_hms("2017-07-05 11:40:00")-ymd_hms("2010-05-24 21:44:00")
chronique.decalage(data, decalage = 2598.581)
}
\keyword{chronique}
