% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.duree.R
\name{chronique.duree}
\alias{chronique.duree}
\title{Chroniques existantes}
\usage{
chronique.duree(Mesures)
}
\arguments{
\item{Data}{Jeu de données issu de la matrice "Mesures" des chroniques}
}
\description{
Cette fonction permet de visualiser les chroniques existantes
}
\examples{
chronique.duree(Mesures)
}
\keyword{chronique}

