% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.figure.interannuelle.R
\name{chronique.figure.interannuelle}
\alias{chronique.figure.interannuelle}
\title{Comparaison interannuelles de chroniques}
\usage{
chronique.figure.interannuelle(
  data = data,
  Titre = "",
  typemesure = c("Thermie", "Thermie barométrique", "Thermie piézométrique",
    "Barométrie", "Piézométrie", "Piézométrie brute", "Piézométrie compensée",
    "Piézométrie calée", "Piézométrie NGF", "Oxygénation", "Hydrologie",
    "Pluviométrie"),
  Ymin = -1,
  Ymax = 30,
  affichagevide = TRUE,
  style = c("boxplot", "violon", "courbes"),
  datedebutanneebiol = "10-01",
  palette = "Dark2",
  save = F,
  projet = NA_character_,
  format = ".png"
)
}
\arguments{
\item{data}{Data.frame contenant a minima une colonne chmes_date et une colonne chmes_valeur}

\item{Titre}{Titre du graphique (vide par défaut)}

\item{typemesure}{Défini le type de données et modifie les légendes en fonction. Ignoré si le champ chmes_typemesure est présent dans data}

\item{Ymin}{Valeur minimale de l'axe des Y (-1 par défaut)}

\item{Ymax}{Valeur maximale de l'axe des Y (aucune par défaut)}

\item{affichagevide}{Si \code{TRUE} (par défaut), fait apparaître les années ne contenant pas de résultats. Si \code{FALSE}, ne fait apparaître que les années contenants des résultats}

\item{style}{En forme de boxplot (par défaut) ou de violon ou de courbes}

\item{datedebutanneebiol}{Date de démarrage de l'année biologique : 10-01 (par défaut - 1er octobre), pour l'affichage sous forme de courbes}

\item{palette}{Nom de la palette à utiliser : \code{Dark2} (par défaut), \code{Spectral}, \code{Set3} ou autre à préciser}

\item{save}{Si \code{FALSE} (par défaut), n'enregistre pas les figures. Si \code{TRUE}, les enregistre.}

\item{projet}{Nom du projet}

\item{format}{Défini le format d'enregistrement (par défaut .png)}
}
\description{
Cette fonction permet de représenter des chroniques de mesures (température, niveaux, etc.) sous forme de comparaison interannuelle
}
\examples{
chronique.figure.interannuelle(data)
data \%>\% chronique.resultats.filtrage() \%>\% chronique.figure.interannuelle()
chronique.figure.interannuelle(data = tableaudonnee, Titre=nom, typemesure = "Barométrie", save=T, format=".png")
chronique.figure.interannuelle(dataVOG4, affichagevide = F, style = "courbes")
}
\keyword{chronique}
