% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.figure.preferendums.R
\name{chronique.figure.preferendums}
\alias{chronique.figure.preferendums}
\title{Vues des préférendums thermiques des especes piscicoles}
\usage{
chronique.figure.preferendums(
  staderecherche = c("Adulte", "Reproduction", "Embryon", "Larve", "Juvénile",
    "Tous stades"),
  tmm30j = 0,
  liste_especes = c("Toutes espèces", "TRF", "CHA", "LPP", "VAI", "LOF", "OBR", "EPI",
    "BLE", "BLN", "CHE", "APR", "GOU", "HOT", "TOX", "BAF", "LOT", "SPI", "VAN", "EPT",
    "BOU", "BRO", "PER", "GAR", "ABL", "CAS", "GRE", "CCO", "SAN", "TAN", "BRB", "BRE",
    "PES", "ROT", "BBG", "PCH", "SIL"),
  titre = "",
  save = F,
  projet = NA_character_,
  format = ".png"
)
}
\arguments{
\item{staderecherche}{Stade de vie recherché : \code{Adulte} (par défaut), \code{Reproduction}, \code{Embryon}, \code{Larve}, \code{Juvénile} ou \code{Tous stades}}

\item{tmm30j}{Valeur de Tmm30j à représenter : 0 par défaut (aucun affichage)}

\item{liste_especes}{Liste des espèces à afficher : toutes par défaut. Les espèces sont affichées dans l'ordre de saisie.}

\item{titre}{Titre du graphique (vide par défaut)}

\item{save}{Si \code{FALSE} (par défault), n'enregistre pas les figures. Si \code{TRUE}, les enregistre.}

\item{projet}{Nom du projet}

\item{format}{Défini le format d'enregistrement (par défaut .png)}
}
\description{
Cette fonction permet de représenter les préférendums thermiques des espèces piscicoles, avec éventuellement la Tmm30j observée
}
\examples{
chronique.figure.preferendums()
chronique.figure.preferendums("Reproduction")
chronique.figure.preferendums("Tous stades", tmm30j = 17.4)
chronique.figure.preferendums("Toutes espèces", tmm30j = 17.4)
chronique.figure.preferendums("Adulte", tmm30j = 17.4, "TRF,CHA,LOG,VAL")
chronique.figure.preferendums("Adulte", tmm30j = 17.4, liste_especes)
chronique.figure.preferendums("Adulte", tmm30j = 17.4, .$chsta_sprep)
}
\keyword{chronique}
\keyword{poissons}
