% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.figure.preferendums.R
\name{chronique.figure.preferendums}
\alias{chronique.figure.preferendums}
\title{Vues des préférendums thermiques des especes piscicoles}
\usage{
chronique.figure.preferendums(
  Titre = "",
  staderecherche = c("Adulte", "Reproduction", "Embryon", "Larve", "Juvénile",
    "Tous stades"),
  tmm30j = 0,
  listeEspeces = c("Toutes espèces", "BAF", "GRE", "PER", "BLN", "BRO", "ABL", "VAN",
    "TOX", "HOT", "PES", "CHE", "GAR", "SIL", "SAN", "GOU", "BOU", "SPI", "BRB", "BRE"),
  save = F,
  projet = NA_character_,
  format = ".png"
)
}
\arguments{
\item{Titre}{Titre du graphique (vide par défaut)}

\item{staderecherche}{Stade de vie recherché : \code{Adulte} (par défaut), \code{Reproduction}, \code{Embryon}, \code{Larve}, \code{Juvénile} ou \code{Tous stades}}

\item{tmm30j}{Valeur de Tmm30j à représenter : 0 par défaut (aucun affichage)}

\item{listeEspeces}{Liste des espèces à afficher : toutes par défaut}

\item{save}{Si \code{FALSE} (par défault), n'enregistre pas les figures. Si \code{TRUE}, les enregistre.}

\item{projet}{Nom du projet}

\item{format}{Défini le format d'enregistrement (par défaut .png)}
}
\description{
Cette fonction permet de représenter les préférendums thermiques des espèces piscicoles, avec éventuellement la Tmm30j observée
}
\examples{
chronique.figure.preferendums()
chronique.figure.preferendums("Reproduction")
chronique.figure.preferendums("Tous stades", tmm30j = 17.4)
chronique.figure.preferendums("Toutes espèces", tmm30j = 17.4)
chronique.figure.preferendums("Adulte", tmm30j = 17.4, c("TRF", "CHA", "VAN", "TOX"))
}
\keyword{chronique}
\keyword{poissons}
