% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.figure.R
\name{chronique.figure}
\alias{chronique.figure}
\title{Représentation de chroniques}
\usage{
chronique.figure(data = data, Titre = "",
  legendeY = "Température (°C)", duree = "Complet", Vmm30j = F,
  Vminmax = T, Ymin = 0, Ymax = NA, save = F, format = ".png")
}
\arguments{
\item{Titre}{Titre du graphique (vide par défaut)}

\item{legendeY}{Défini le texte de la légende de l'axe Y (Température (C))}

\item{duree}{Si \code{Complet} (par défault), affichage de l'année complète.  Si \code{Relatif}, affichage uniquement de la période concernée.}

\item{Vmm30j}{Si \code{FALSE} (par défault), n'affiche pas les
Vmm30j.  Si \code{TRUE}, les affiche.}

\item{Vminmax}{Si \code{TRUE} (par défault), affiche pas les
valeurs journalières minimales et maximales.  Si \code{FALSE}, ne les affiche pas.}

\item{Ymin}{Valeur minimale de l'axe des Y (0 par défaut)}

\item{Ymax}{Valeur maximale de l'axe des Y (aucune par défaut)}

\item{save}{Si \code{FALSE} (par défault), n'enregistre pas les
figures.  Si \code{TRUE}, les enregistre.}

\item{format}{Défini le format d'enregistrement (par défaut .png)}

\item{thermie}{Data.frame contenant a minima une colonne Date et une colonne Valeur}
}
\description{
Cette fonction permet de représenter des chroniques de mesures (température, niveaux, etc.)
}
\examples{
chronique.figure(data)
chronique.figure(data = mesdonnees, legendeY = "Température (°C)", duree = "Complet")
chronique.figure(data = tableaudonnee, Titre=nom, legendeY = "Température (°C)", save=T, format=".png")
}
\keyword{chronique}
