% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.figure.R
\name{chronique.figure}
\alias{chronique.figure}
\title{Représentation de chroniques}
\usage{
chronique.figure(
  data = data,
  Titre = "",
  typemesure = c("Thermie", "Thermie barométrique", "Thermie piézométrique",
    "Barométrie", "Piézométrie", "Piézométrie brute", "Piézométrie compensée",
    "Piézométrie calée", "Piézométrie NGF", "Oxygénation", "Hydrologie",
    "Pluviométrie"),
  duree = c("Complet", "Relatif"),
  complement = FALSE,
  Vmm30j = F,
  Vminmax = T,
  Ymin = -1,
  Ymax = NA,
  save = F,
  projet = NA_character_,
  format = ".png"
)
}
\arguments{
\item{data}{Data.frame contenant a minima une colonne chmes_date et une colonne chmes_valeur}

\item{Titre}{Titre du graphique (vide par défaut)}

\item{typemesure}{Ignoré si le champ chmes_typemesure est présent dans data. Défini le type de données et modifie les légendes en fonction (\code{Thermie}, \code{Thermie barométrique}, \code{Thermie piézométrique}, \code{Barométrie}, \code{Piézométrie}, \code{Piézométrie brute}, LI70DU6rWs1yg9RLXlOh1BoUtpBg3pax-7\code{Piézométrie NGF}, \code{Oxygénation}, \code{Hydrologie}, \code{Pluviométrie}.}

\item{duree}{Si \code{Complet} (par défault), affichage de l'année complète.  Si \code{Relatif}, affichage uniquement de la période concernée.}

\item{complement}{Si \code{TRUE}, complément de la chronique avec les données manquantes (\code{FALSE} par défaut)}

\item{Vmm30j}{Si \code{FALSE} (par défault), n'affiche pas les
Vmm30j.  Si \code{TRUE}, les affiche.}

\item{Vminmax}{Si \code{TRUE} (par défault), affiche pas les
valeurs journalières minimales et maximales.  Si \code{FALSE}, ne les affiche pas.}

\item{Ymin}{Valeur minimale de l'axe des Y (-1 par défaut)}

\item{Ymax}{Valeur maximale de l'axe des Y (aucune par défaut)}

\item{save}{Si \code{FALSE} (par défault), n'enregistre pas les
figures.  Si \code{TRUE}, les enregistre.}

\item{projet}{Nom du projet}

\item{format}{Défini le format d'enregistrement (par défaut .png)}
}
\description{
Cette fonction permet de représenter des chroniques de mesures (température, niveaux, etc.)
}
\examples{
chronique.figure(data)
chronique.figure(data = mesdonnees, typemesure = "Thermie", duree = "Complet")
chronique.figure(data = tableaudonnee, Titre=nom, typemesure = "Barométrie", save=T, format=".png")
}
\keyword{chronique}
