% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.hydroreel.R
\name{chronique.hydroreel}
\alias{chronique.hydroreel}
\title{Collecte de la dernière valeur de débit transmise}
\usage{
chronique.hydroreel(Station)
}
\arguments{
\item{station}{Numéro de la station}
}
\description{
Cette fonction permet d'extraire depuis Hydroréel la dernière valeur mesurée de débit d'une station donnée
}
\examples{
chronique.hydroreel(Station)
chronique.hydroreel(433)
# Extraction des mesures à partir de http://www.rdbrmc.com/hydroreel2/listestation.php?dep=39
}
\keyword{chronique}
