% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.mesures.R
\name{chronique.mesures}
\alias{chronique.mesures}
\title{Extraction de données brutes de chronique}
\usage{
chronique.mesures(CodeStation = "DRO6-8", Type = c("Thermie",
  "Piézométrie", "Hydrologie", "O2", "Pluviométrie"), annee = "2013")
}
\description{
Permet d'extraire les données de chronique d'une station
}
\examples{
chronique.mesures("HER0-6", "Thermie") 
chronique.mesures("DRO6-8", "Thermie", "2013")
}
\keyword{chronique}

