% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.ouverture.R
\name{chronique.ouverture}
\alias{chronique.ouverture}
\title{Ouverture de fichiers de chroniques}
\usage{
chronique.ouverture(
  Type = c("Mesures", "Suivis", "Stations", "Commentaires", "Capteurs"),
  typemesure = c("Thermie", "Thermie barométrique", "Thermie piézométrique",
    "Barométrie", "Piézométrie", "Piézométrie brute", "Piézométrie compensée",
    "Piézométrie calée", "Piézométrie NGF", "Oxygénation", "Hydrologie",
    "Pluviométrie", "Télétransmission"),
  Localisation = NA_character_,
  feuille = 1,
  skipvalue = 9,
  nbcolonnes = 2,
  separateur_colonnes = c(";", ","),
  separateur_decimales = c(",", "."),
  typefichier = c(".csv", "excel", ".ods", "Interne"),
  typedate = c("ymd", "dmy", "mdy", "dmy_hms", "dmy_hm", "mdy_hms", "mdy_hm", "ymd_hms"),
  typecapteur = c("Non précisé", "Hobo", "Diver", "VuSitu", "miniDOTindividuel",
    "miniDOTregroupe", "EDF", "RuggedTROLL", "Aquaread", "WiSens", "VuLink", "Cube"),
  nomfichier = F,
  formatmacmasalmo = F
)
}
\arguments{
\item{Type}{Type de données en entrée (Mesures, Suivis, Stations, Commentaires, Capteurs)}

\item{typemesure}{Défini le type de données (Thermie, Piézométrie, etc.)}

\item{Localisation}{Localisation relative du fichier (à partir de /NAS-DATA/)}

\item{feuille}{Feuille où lire les données dans le cas de l'ouverture d'un fichier excel : \code{1} (par défaut)}

\item{skipvalue}{Nombre de lignes à sauter en début de fichier (1 par défaut pour les mesures)}

\item{nbcolonnes}{Nombre de colonnes concernées}

\item{separateur_colonnes}{Séparateur de colonnes : \code{;} (par défaut), \code{,}}

\item{separateur_decimales}{Séparateur de colonnes : \code{,} (par défaut), \code{.}}

\item{typefichier}{Type de fichier : \code{.csv} (par défaut), \code{excel}, \code{.ods} ou \code{Interne}}

\item{typedate}{Format des dates pour les mesures (ymd par défaut, dmy, mdy, dmy_hms, dmy_hm, mdy_hms, mdy_hm ou ymd_hms)}

\item{typecapteur}{Type de capteur, pour les piézomètres ou les capteurs O2 (\code{Non précisé} par défaut)}

\item{nomfichier}{Ajoute une colonne \code{nom_fichier} avec le nom du fichier traité : \code{FALSE} (par défault)}

\item{formatmacmasalmo}{Format d'entrée nécessaire à MacmaSalmo (Heure puis date puis valeur) : \code{FALSE} (par défault)}
}
\description{
Cette fonction permet d'ouvrir de manière semi-automatisée des fichiers de chroniques
}
\examples{
DataToAdd <- chronique.ouverture("Mesures", "Thermie", "/NAS-DATA/Chroniques/Saisie_JB/2018-12-17_Suivi_Ain_amont_FJPPMA_été_2018/10880567.txt")
DataToAdd <- chronique.ouverture("Stations", "Thermie", "/NAS-DATA/Chroniques/Saisie_JB/2019-01-11_PDPG/Stations sondesPDPG.xlsx")
chronique.ouverture("Suivi", "Thermie", "/NAS-DATA/Chroniques/Saisie_JB/2019-01-18_PDPG/Saisie_cahier_terrain_PDPG_V2.xlsx")
}
\keyword{chronique}
