% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.periode.R
\name{chronique.periode}
\alias{chronique.periode}
\title{Filtrage d'une chronique}
\usage{
chronique.periode(x = Data, deb = "", fin = "")
}
\arguments{
\item{Début}{Date de début (2015-02-01 14:00:00)}

\item{Fin}{Date de fin (2015-02-04 14:00:00)}
}
\description{
Cette fonction permet de filtrer la période d'intérêt. La valeur de début est l'heure sélectionnée +1, la valeur de fin est strictement inférieure à celle indiquée
}
\examples{
chronique.periode(Data, "2015-02-04 11:00:00", )
chronique.periode(Data, ,"2015-02-04 14:00:00")
chronique.periode(Data, "2015-02-04 11:00:00", "2015-02-04 14:00:00")
}
\keyword{chronique}
