% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.regroupement.R
\name{chronique.regroupement}
\alias{chronique.regroupement}
\title{Regroupement de chroniques}
\usage{
chronique.regroupement(
  data = data,
  stations = NA_character_,
  formatcle = c("STU", "MSTU"),
  projet = NA_character_,
  export = FALSE
)
}
\arguments{
\item{data}{Data.frame contenant a minima une colonne chmes_date, une colonne chmes_heure, une colonne chmes_coderhj, une colonne chmes_typemesure et une colonne chmes_valeur}

\item{stations}{Dataframe de stations, avec au moins une colonne chsta_coderhj et une colonne chsta_distancesource. Permet le classement des stations par ordre de distance à la source}

\item{formatcle}{Format de la clé d'identification (\code{STU} par défaut)}

\item{projet}{Nom du projet}

\item{export}{Si \code{TRUE}, exporte les résultats (\code{FALSE} par défaut)}
}
\description{
Cette fonction permet de regrouper des chroniques de mesures (température, niveaux, etc.) dans un tableau au format large, avec une colonne date, une colonne heure puis des colonnes avec les différentes stations/paramètres
}
\examples{
DataTravail <- chronique.regroupement(data)
DataTravail <- chronique.regroupement(data, formatcle = "MSTU")
}
\keyword{chronique}
