% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.stations.R
\name{chronique.stations}
\alias{chronique.stations}
\title{Listage des stations de chroniques}
\usage{
chronique.stations(x = "CD39", Recherche = c("MO", "Milieu", "Bassin",
  "Sous-bassin", "Commune", "Département"))
}
\arguments{
\item{Recherche}{Type de données recherchées}

\item{Nom}{Nom recherché}
}
\description{
Cette fonction permet de lister les stations de la BDD Chroniques
}
\examples{
chronique.stations("CD39","MO")
chronique.stations("Suran","Milieu")
chronique.stations("Ain","Bassin")
chronique.stations("Serpentine","Sous-bassin")
chronique.stations("Lons-le-Saunier","Commune")
chronique.stations("39","Département")
}
\keyword{stations}
