% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.traitement.R
\name{chronique.traitement}
\alias{chronique.traitement}
\title{Exécuter le traitement de données de chroniques}
\usage{
chronique.traitement(data = data)
}
\arguments{
\item{data}{Data.frame issu de chronique.mesures, pouvant contenir différentes stations}
}
\description{
Permet d'exécuter plusieurs opérations de traitement des données de chronique d'une station
}
\examples{
chronique.traitement(data)
DataTravail <- chronique.mesures("SUR0-9", "Thermie") \%>\% chronique.traitement()
DataTravail <- tbl(db,"Mesures") \%>\% filter(CodeRDT == "DOU393-2" | CodeRDT == "DOU394-5") \%>\% collect() \%>\% chronique.traitement()
}
\keyword{chronique}
