% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronique.validation.R
\name{chronique.validation}
\alias{chronique.validation}
\title{Validation des chroniques}
\usage{
chronique.validation(data)
}
\arguments{
\item{data}{Chronique à valider}
}
\description{
Cette fonction permet d'obtenir des éléments afin de valider les chroniques
}
\examples{
chronique.validation(data)
}

