% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.annee.biologique.R
\name{formatage.annee.biologique}
\alias{formatage.annee.biologique}
\title{Reformatage de dates en année biologique}
\arguments{
\item{data}{Jeu de données contenant une colonne chmes_date ou chsvi_date au format Date}
}
\description{
Reformate les dates selon l'année biologique (en rapport au 1 octobre de chaque année) dans une colonne chmes_anneebiol
}
\examples{
formatage.annee.biologique(data)
}
\keyword{donnees}
