% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.date.heure.R
\name{formatage.date.heure}
\alias{formatage.date.heure}
\title{Reformatage d'une colonne time en deux colonnes de dates et heures}
\usage{
formatage.date.heure(data)
}
\arguments{
\item{data}{Jeu de données contenant une colonne \code{time}}
}
\description{
Reformate une colonne time en deux colonnes de dates et heures
}
\examples{
formatage.date.heure(data)
}
\keyword{donnees}
