% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatage.ecosysteme.R
\name{formatage.ecosysteme}
\alias{formatage.ecosysteme}
\title{Mise en acronymes des écosystèmes ou inversement}
\arguments{
\item{acronymes}{Dataframe contenant les données à transformer}

\item{Operation}{Type de transformation que l'on souhaite réaliser}
}
\description{
Cette fonction permet de mettre en acronymes des noms d'écosystèmes ou de transformer des acronymes en noms des écosystèmes
}
\examples{
formatage.ecosysteme(data$Station, Operation = "Simplification")
PC$pcmes_coderhj <- formatage.ecosysteme(PC$pcmes_coderhj, Operation = "Simplification")
data \%>\% stations.coderhj(DistSource = F) \%>\% formatage.ecosysteme(Operation = "Expansion")
}
