% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.captures.R
\name{MI.captures}
\alias{MI.captures}
\title{Extraction des données des captures de MI}
\usage{
MI.captures(station = "VAL37-4", date = "2013-07-22")
}
\description{
Récupère les données de captures d'une opération de suivi MI
}
\examples{
MI.captures("VAL37-4", "2013-07-22")
}
\keyword{donnees}

