% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.nettoyage.R
\name{MI.nettoyage}
\alias{MI.nettoyage}
\title{Nettoyage des données de MI}
\usage{
MI.nettoyage(data)
}
\arguments{
\item{data}{Chronique à valider}
}
\description{
Cette fonction permet de corriger des codes erronés des données de MI
}
\examples{
MI.nettoyage(data)
}
\keyword{MI}
