% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.figure.parametres.R
\name{PC.figure.parametres}
\alias{PC.figure.parametres}
\title{Paramétrisation des vues de physico-chimie}
\usage{
PC.figure.parametres(data = data, typefigure = c("valeurs"))
}
\arguments{
\item{data}{Data.frame issu de la fonction PC.contexte()}

\item{typefigure}{Type de figure qui est à paramétriser : \code{valeurs} (par défault)}
}
\description{
Cette fonction permet de sélectionner automatiquement les paramètres de légende des représentations graphiques de physico-chimie à partir des données issues de PC.contexte
}
\examples{
data \%>\% PC.contexte() \%>\% PC.figure.parametres()

parametres <- data \%>\% PC.contexte() \%>\% PC.figure.parametres()
legende_y <- parametres$legende_y
legende_titre <- parametres$legende_titre
typemesure_titre_sortie <- parametres$typemesure_titre_sortie
}
\keyword{physico-chimie}
