% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.figure.R
\name{PC.figure}
\alias{PC.figure}
\title{Représentation de résultats de PC}
\arguments{
\item{PC}{Jeu de données issu de la BDD PC}
}
\description{
Cette fonction permet de représenter des valeurs physico-chimiques
}
\examples{
PC.figure(PC)
PC \%>\% PC.figure()
spNitrates <- ChimieOrg \%>\% filter(ParametreNom == Parametre[3]) \%>\% PC.figure()
}
