% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.hubeau.R
\name{PC.hubeau}
\alias{PC.hubeau}
\title{Collecte de données de physico-chimie depuis Hub'eau}
\usage{
PC.hubeau(
  codesie = NA_character_,
  parametresandre = NA_character_,
  supportsandre = c("3", "6"),
  coderemarque = c("1"),
  qualificationsandre = c("1")
)
}
\arguments{
\item{codesie}{Code SANDRE de la station concernée}

\item{parametresandre}{Code SANDRE du paramètre concerné}

\item{supportsandre}{Code SANDRE du support concerné : \code{3} - Eau (par défaut), \code{6} - Sédiments}

\item{coderemarque}{Code SANDRE du code remarque concerné : \code{1} - Résultat > seuil de quantification et < au seuil de saturation (par défaut)}

\item{qualificationsandre}{Code SANDRE de la qualification concernée : \code{1} - Correcte (par défaut)}
}
\description{
Cette fonction permet de collecter de données de physico-chimie depuis Hub'eau
}
\examples{
PC.hubeau(codesie, parametresandre)
PC.hubeau("06400085", "1340")
PC.hubeau("06400085", "1386", supportsandre = "6")
}
