% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.lac.profil.R
\name{PC.lac.profil}
\alias{PC.lac.profil}
\title{Création de profils verticaux lacustres}
\usage{
PC.lac.profil(
  PC,
  param = c("O2mg", "O2pourc", "cond", "ph", "temp", "redox", "chlorophylles",
    "phycocyanines", "secchi")
)
}
\arguments{
\item{param}{Paramètre physico-chimique à représenter (O2mg, O2pourc, cond, ph, temp, redox, chlorophylles, phycocyanines, secchi)}
}
\description{
Cette fonction permet de créer des profils graphiques de paramètres physico-chimiques
}
\examples{
PC.lac.profil(PC, param = "O2mg")
PC \%>\% PC.lac.profil(param = "redox")
}
\keyword{physico-chimie}
