% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.saturationO2.R
\name{PC.saturationO2}
\alias{PC.saturationO2}
\title{Saturation en oxygène}
\arguments{
\item{data}{Jeu de données (contenant une colonne Thermie et une colonne Oxygenation)}
}
\description{
Cette fonction permet de calculer des saturations en oxygène à partir des chroniques de mesures de concentration et de température de l'eau
}
\examples{
PC.saturationO2(data)
# Calcul à partir de https://projects.ncsu.edu/cals/course/zo419/oxygen.html
}
\keyword{Physico-chimie}
