% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.CAAvue.R
\name{poissons.CAAvue}
\alias{poissons.CAAvue}
\title{Affichage de graphes piscicoles}
\description{
Cette fonction permet de créer des comparaisons entre les espèces attendues et observées. Le jeu de données doit contenir un champ codeespece, un champ CA et un champ date
}
\examples{
poissons.CAAvue(data)
poissons.CAAvue(data, save=T)
poissons.CAA("MAD6-2", 5) \%>\% poissons.CAAvue()
poissons.CAA("MAD6-2", 5) \%>\% poissons.CAAvue(save = T)
}
\keyword{poissons}
