% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.captures.R
\name{poissons.captures}
\alias{poissons.captures}
\title{Extraction des données de captures de poissons pour une opération}
\usage{
poissons.captures(station = "SOR10-2", date = "2015-05-19")
}
\description{
Récupère les données de captures de poissons d'une opération à partir de Maxifish dans un dataframe
}
\examples{
poissons.captures("SOR10-2", "2015-05-19")
}
\keyword{donnees}

