% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.especes.R
\name{poissons.especes}
\alias{poissons.especes}
\title{Extraction des données de référence des espèces}
\arguments{
\item{Sortie}{Format de la sortie}

\item{Couleurs}{Présence de couleurs dans le formatage du statut des espèces (\code{FALSE} par défault)}
}
\description{
Récupère l'ensemble des données de référence des espèces de Multifish
}
\examples{
poissons.especes()
poissons.especes(Sortie = "Simple")
poissons.especes("Propre")
poissons.especes("Propre", Couleurs = T)
}
\keyword{donnees}
