% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.expertise.R
\name{poissons.expertise}
\alias{poissons.expertise}
\title{Exportation des avis à dire d'expert}
\arguments{
\item{ListeOperations}{Dataframe contenant un colonne "Station" avec le code de la station (RHJ) et une colonne "Date"}
}
\description{
Exportation des avis à dire d'expert
}
\examples{
poissons.expertise(data.frame(Station = "SOR10-2", Date = "2012-11-03"))
}
\keyword{poissons}
