% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.IPR.R
\name{poissons.IPR}
\alias{poissons.IPR}
\title{Exportation des IPR}
\usage{
poissons.IPR(station = "LEU15-2", date = "2013-10-22", expertise = TRUE)
}
\arguments{
\item{station}{Code de la station}

\item{date}{Date de la pêche}

\item{expertise}{\code{TRUE} par défault}
}
\description{
Cette fonction permet d'exporter les résultats IPR de pêche
}
\examples{
poissons.IPR("SOR10-2", "2015-05-19")
}
\keyword{poissons}

