% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.NTT.R
\name{poissons.NTT}
\alias{poissons.NTT}
\title{Calcul des NTT}
\usage{
poissons.NTT(station)
}
\description{
Cette fonction permet de calculer les Niveaux Typologiques Théoriques à partir d'un dataframe issu de poissons.stations()
}
\examples{
poissons.NTT(station)
}
\keyword{poissons}

