% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.operations.R
\name{poissons.operations}
\alias{poissons.operations}
\title{Listing des opérations pour une station}
\usage{
poissons.operations(ListeStations = data.frame(Nom = character(0)),
  CodeOperation = FALSE, Sortie = c("Simple", "Propre", "Complet"))
}
\arguments{
\item{ListeStations}{Dataframe contenant un colonne "Nom" avec le code de la station (RHJ)}

\item{CodeOperation}{Affichage du CodeOperation - \code{FALSE} (par défault)}

\item{Sortie}{Forme du dataframe de sortie - \code{Simple} (par défault), \code{Propre} (format diffusable, avec stations) ou \code{Complet} (tous les champs)}
}
\description{
Cette fonction permet de lister les différentes opérations pour une liste donnée de stations
}
\examples{
poissons.operations()
poissons.operations(CodeOperation = TRUE)
poissons.operations(data.frame(Nom = "SOR10-2"))
poissons.operations(data.frame(Nom = "SOR10-2"), CodeOperation = TRUE)
poissons.operations(listeStations, Sortie = Propre)
poissons.operations(data.frame(Nom = "SOR10-2"), Sortie = Complet)
}
\keyword{poissons}
