% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.operations.R
\name{poissons.operations}
\alias{poissons.operations}
\title{Listing des opérations pour une station}
\usage{
poissons.operations(station = "")
}
\arguments{
\item{station}{Code de la station}
}
\description{
Cette fonction permet de lister les différentes opérations pour une station donnée
}
\examples{
poissons.operations()
poissons.operations("SOR10-2")
}
\keyword{poissons}

