% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.poids.R
\name{poissons.poids}
\alias{poissons.poids}
\title{Calcul des poids attendus en fonction de la taille pour un jeu de captures}
\usage{
poissons.poids(Captures = Captures)
}
\description{
Calcule, à partir d'une jeu de données de captures, les poids attendus pour des tailles données
}
\examples{
poissons.poids(Captures)
}
\keyword{donnees}

