% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.resultats.R
\name{poissons.resultats}
\alias{poissons.resultats}
\title{Extraction des données des résultats d'une opération}
\description{
Récupère les données de résultats d'une opération dans Maxifish pour les mettre dans un dataframe
}
\examples{
poissons.resultats("SOR10-2", "2015-05-19")
}
\keyword{donnees}
