% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.synthese.R
\name{poissons.synthese}
\alias{poissons.synthese}
\title{Synthèse des données piscicoles}
\usage{
poissons.synthese(Synthese = "Présence", Echelle = c("Commune",
  "ComCom", "ContextePDPG", "Département"), Territoire = "39")
}
\arguments{
\item{Synthese}{Type de synthèse. Présence par défaut}

\item{Echelle}{Échelle spatiale de la synthèse (commune, communauté de communes, contexte de PDPG ou département)}

\item{Territoire}{Territoire concerné : code insee ou code contexte}
}
\description{
Cette fonction permet de réaliser différentes synthèses des enjeux piscicoles à différentes échelles spatiales (commune, communauté de communes, contexte de PDPG ou département)
}
\examples{
poissons.synthese("Présence", "ComCom", "200071595")
}
