% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons.traitement.R
\name{poissons.traitement}
\alias{poissons.traitement}
\title{Traitement semi-automatisé de données piscicoles}
\arguments{
\item{data}{Jeu de données en entrée (captures issues de poissons.captures)}

\item{export}{\code{FALSE} par défault. Permet d'exporter les données}

\item{commentaires}{\code{FALSE} par défault. Permet d'extraire le commentaire associé aux opérations}
}
\description{
Cette fonction permet de traiter de manière semi-automatisée des données piscicoles (histogrammes, graphiques CAA/CAR, etc.)
}
\examples{
poissons.traitement(data)
poissons.traitement(data, export=T)
poissons.captures("AIN18-4") \%>\% poissons.traitement(export=T)
}
\keyword{poissons}
