% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissons_synthese_CAR.R
\name{poissons_synthese_CAR}
\alias{poissons_synthese_CAR}
\title{Calcul et regroupement des CAR par espèce par station}
\usage{
poissons_synthese_CAR(poissons)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
poissons_synthese_CAR()
}
\keyword{cats}
