% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profil.lac.R
\name{profil.lac}
\alias{profil.lac}
\title{Création de profils verticaux lacustres}
\usage{
profil.lac(PC, param = "O2mg")
}
\arguments{
\item{O2mg}{Pour réaliser un profil de concentration en oxygène dissous}

\item{O2pourc}{Pour réaliser un profil de saturation en oxygène dissous}

\item{Cond}{Pour réaliser un profil de conductivité}

\item{ph}{Pour réaliser un profil de pH}

\item{temp}{Pour réaliser un profil de température}
}
\description{
Cette fonction permet de créer des profils graphiques de paramètres physico-chimiques
}
\examples{
profil.lac(PC,param="O2mg")
profil.lac(PC,param="O2pourc")
profil.lac(PC,param="Cond")
profil.lac(PC,param="ph")
#' profil.lac(PC,param="temp")
}
