% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profil.lac.R
\name{profil.lac}
\alias{profil.lac}
\title{Création de profils verticaux lacustres}
\usage{
profil.lac(PC, param = "O2mg")
}
\arguments{
\item{O2mg}{}

\item{Cond}{}
}
\description{
Cette fonction permet de créer des profils graphiques de paramètres physico-chimiques
}
\examples{
profil.lac(PC,param="O2mg")
profil.lac(PC,param="O2pourc")
profil.lac(PC,param="Cond")
profil.lac(PC,param="ph")
#' profil.lac(PC,param="temp")
}

