% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG.conversion.dec2l93.R
\name{SIG.conversion.dec2l93}
\alias{SIG.conversion.dec2l93}
\title{Conversion de coordonnées en degrés décimaux vers Lambert 93}
\usage{
SIG.conversion.dec2l93(coord_x, coord_y)
}
\arguments{
\item{coord_x}{Coordonnée x à convertir}

\item{coord_y}{Coordonnée y à convertir}
}
\description{
Cette fonction permet de convertir des coordonnées en egrés décimaux vers du Lambert 93
}
\examples{
SIG.conversion.dec2l93(46.806626, 5.972321)
}
\keyword{SIG}
