% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG.conversion.R, R/SIG.conversion.l93dec2.R
\name{SIG.conversion.l93dec2}
\alias{SIG.conversion.l93dec2}
\alias{SIG.conversion}
\title{Conversion de coordonnées}
\usage{
SIG.conversion(coord_x, coord_y, projection_source = NA, projection_cible = NA)

SIG.conversion.l93dec2(coord_x = NA, coord_y = NA)
}
\arguments{
\item{coord_x}{Coordonnée x à convertir}

\item{coord_y}{Coordonnée y à convertir}

\item{projection_source}{Projection initiale}

\item{projection_cible}{Projection cible}
}
\description{
Cette fonction permet de convertir des coordonnées d'une projection vers une autre

Cette fonction permet de convertir des coordonnées en du Lambert 93 (2154) vers degrés décimaux (4326)
}
\examples{
SIG.conversion(888227, 6655627, 2154, 4326)
SIG.conversion.l93dec2(888227, 6655627)
}
\keyword{SIG}
