% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG.export.R
\name{SIG.export}
\alias{SIG.export}
\title{Export de données en shp}
\arguments{
\item{sata}{Dataframe à exporter}

\item{nomfichier}{Nom du fichier de sortie}
}
\description{
Cette fonction permet d'exporter un dataframe (avec deux colonnes X et Y) ou un sf en Lambert 93 dans un shapefile/geojson/kml/excel
}
\examples{
SIG.export(Stations, "Export_stations")
DataToAdd \%>\% st_as_sf(coords = c("XLong", "YLat"), crs = 4269) \%>\% st_transform(2154) \%>\% SIG.export("GCL_Bathy_points_bruts")
}
\keyword{poissons}
\keyword{stations}
