% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG.export.R
\name{SIG.export}
\alias{SIG.export}
\title{Export de données en shp}
\usage{
SIG.export(
  data = Stations,
  nomfichier = glue("{today()}_Export_stations"),
  shp = TRUE,
  kml = TRUE,
  geojson = TRUE,
  excel = TRUE
)
}
\arguments{
\item{data}{Dataframe à exporter}

\item{nomfichier}{Nom du fichier de sortie}

\item{shp}{Export au format shapefile (\code{true} par défaut)}

\item{kml}{Export au format kml (\code{true} par défaut)}

\item{geojson}{Export au format geojson (\code{true} par défaut)}

\item{excel}{Export au format excel (\code{true} par défaut)}
}
\description{
Cette fonction permet d'exporter un dataframe (avec deux colonnes X et Y) ou un sf en Lambert 93 dans un shapefile/geojson/kml/excel
}
\examples{
SIG.export(Stations, glue("{today()}_Export_stations"))
Stations \%>\% SIG.export("Stations", shp = F, kml = F, excel = F)
DataToAdd \%>\% st_as_sf(coords = c("XLong", "YLat"), crs = 4269) \%>\% st_transform(2154) \%>\% SIG.export("GCL_Bathy_points_bruts")
}
\keyword{poissons}
\keyword{stations}
