% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG.exportSHP.R
\name{SIG.exportSHP}
\alias{SIG.exportSHP}
\title{Export de données en shp}
\usage{
SIG.exportSHP(Data = Stations, nomfichier = "Export_stations.shp")
}
\arguments{
\item{Data}{Dataframe à exporter}

\item{nomfichier}{Nom du fichier de sortie}
}
\description{
Cette fonction permet d'exporter un dataframe (avec deux colonnes X et Y) dans un shapefile en Lambert 93
}
\examples{
SIG.exportSHP(Stations, "Export_stations.shp")
}
\keyword{poissons}
\keyword{stations}
