% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIG.exportSHP.R
\name{SIG.exportSHP}
\alias{SIG.exportSHP}
\title{Export de données en shp}
\usage{
SIG.exportSHP(Data = Stations, a = Stations$XL93, b = Stations$YL93,
  nomfichier = "Export_stations")
}
\arguments{
\item{Data}{Dataframe à exporter}

\item{nomfichier}{Nom du fichier de sortie}

\item{XLambert}{Colonne des X}

\item{YLambert}{Colonne des Y}
}
\description{
Cette fonction permet d'exporter un shapefile en Lambert 93
}
\examples{
SIG.exportSHP(Stations, "Stations$XL93", "Stations$YL93", "Export_stations")
}
\keyword{poissons}
\keyword{stations}

