% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermie.mesures.R
\name{thermie.mesures}
\alias{thermie.mesures}
\title{Extraction de données de thermie}
\usage{
thermie.mesures(station = "DRO6-8", annee = "2013")
}
\description{
Permet d'extraire les données de thermie d'une station pour une année biologique
}
\examples{
thermie.mesures("DRO6-8", "2013")
}
\keyword{donnees}

