% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topographie.figure.profil.R
\name{topographie.figure.profil}
\alias{topographie.figure.profil}
\title{Vue d'un profil le long d'un transect}
\usage{
topographie.figure.profil(
  data_profil_complet = NA_character_,
  data_points_remarquables = NA_character_,
  altitude_ligne_deau = 0,
  points_remarquables = FALSE,
  points_remarquables_etiquette = FALSE,
  save = FALSE,
  save_name = NA_character_
)
}
\arguments{
\item{data_profil_complet}{Jeu de données en sortie de \code{topographie.profil} ou de \code{topographie.profil.raster}}

\item{data_points_remarquables}{Jeu de données en sortie de \code{topographie.profil} ou de \code{topographie.profil.raster}}

\item{altitude_ligne_deau}{Altitude de la ligne d'eau à afficher. Si \code{0} (par défaut), pas d'affichage}

\item{points_remarquables}{Affichage des points remarquables projetés le long du transect : \code{TRUE} (par défaut) ou \code{FALSE}}

\item{points_remarquables_etiquette}{Affichage des étiquettes des points remarquables : \code{TRUE} (par défaut) ou \code{FALSE}}

\item{save}{Si \code{FALSE} (par défaut), n'enregistre pas les figures. Si \code{TRUE}, les enregistre.}

\item{save_name}{Nom de fichier à utiliser pour l'enregistrement. \code{"{today()}_Vue_profil.png"} si vide.}
}
\description{
Cette fonction permet de représenter un profil (topographique par exemple) d'un transect
}
\examples{
data \%>\% topographie.figure.profil()
topographie.figure.profil(data, points_remarquables_etiquette = F, points_remarquables = F)
data \%>\% topographie.figure.profil(save = T)
data \%>\% topographie.figure.profil(save = T, save_name = "transect_2")
topographie.figure.profil(raster = mon_raster_regroupe, transect = StationsTransect, points_projetes = Stations, points_projetes_position = "intermediaire") \%>\% topographie.figure.profil()
}
\keyword{topographie}
